% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/varray_helper_functions.R
\name{Atocon}
\alias{Atocon}
\alias{contoA}
\title{Convert Vine Array to "Convenient" Array}
\usage{
Atocon(A)

contoA(Acon)
}
\arguments{
\item{A}{Vine array}

\item{Acon}{A convenient vine array}
}
\value{
A vine array or convenient vine array
}
\description{
A "convenient" array is achieved by moving the variables in a
(possibly truncated) vine array to the top row.
}
\details{
\code{Atocon} converts a vine array to a convenient vine array.
\code{contoA} converts a convenient vine array to a vine array.

These converters are intended to be used internally.
}
\note{
This form of array is "convenient" because truncated vines can
be respresented by truncating the bottom rows of the matrix.
}

