% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addlayer.R
\name{addlayer.rvine}
\alias{addlayer.rvine}
\title{Add Layers to a Regular Vine}
\usage{
\method{addlayer}{rvine}(rv, a, cops, cpars)
}
\arguments{
\item{rv}{Object of type "rvine".}

\item{a}{The vine array columns to add. Could be a vector if only one
variable is being added; or, could be a list of vectors for each variable
being added. Or it could be the matrix itself.}

\item{cops}{The copula models for each edge. Input as in \code{a}.}

\item{cpars}{The copula parameters for each copula model. Input as in \code{a}.}
}
\value{
Object of type 'rvine' with the new layer(s) added.
}
\description{
Add Layers to a Regular Vine
}

