% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cop-bb1_supp.R
\name{rbb1r}
\alias{dbb1u}
\alias{dbb1v}
\alias{logdbb1u}
\alias{logdbb1v}
\alias{pbb1u}
\alias{pbb1v}
\alias{pcondbb1u}
\alias{pcondbb1v}
\alias{qcondbb1u}
\alias{qcondbb1v}
\alias{rbb1r}
\alias{rbb1u}
\alias{rbb1v}
\title{BB1 copula: supplemental functions}
\usage{
rbb1r(n, cpar)

pbb1u(u, v, cpar)

dbb1u(u, v, cpar)

logdbb1u(u, v, cpar)

rbb1u(n, cpar)

pcondbb1u(v, u, cpar)

qcondbb1u(p, u, cpar)

pbb1v(u, v, cpar)

dbb1v(u, v, cpar)

logdbb1v(u, v, cpar)

rbb1v(n, cpar)

pcondbb1v(v, u, cpar)

qcondbb1v(p, u, cpar)
}
\description{
Appended "u" means a horizontal flip (i.e. switching u to 1-u).
Appended "v" means a vertical flip (i.e. switching v to 1-v).
Appended "r" means reflection/survival copula
}

