% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb1rsk-qcond.R, R/bb1rsk.R
\name{qcondbb1rsk21}
\alias{dbb1rsk}
\alias{pbb1rsk}
\alias{pcondbb1rsk}
\alias{pcondbb1rsk12}
\alias{pcondbb1rsk21}
\alias{qcondbb1rsk}
\alias{qcondbb1rsk12}
\alias{qcondbb1rsk21}
\title{Skew BB1r Copula Family}
\usage{
qcondbb1rsk21(p, u, cparsk, eps = 1e-08, mxiter = 30, iprint = F)

qcondbb1rsk12(p, v, cparsk, eps = 1e-08, mxiter = 30, iprint = F)

qcondbb1rsk(p, u, cparsk, eps = 1e-08, mxiter = 30, iprint = F)

pbb1rsk(u, v, cparsk)

pcondbb1rsk21(v, u, cparsk)

pcondbb1rsk(v, u, cparsk)

pcondbb1rsk12(u, v, cparsk)

dbb1rsk(u, v, cparsk)
}
\arguments{
\item{cparsk}{copula parameter vector, third component in [0,1]}

\item{eps}{tolerance for convergence}

\item{mxiter}{maximum number of Newton-Raphson iterations}
}
\description{
Adds a skew parameter alpha in [0,1]. Distribution function is found by
\code{pbb1r(u/u^alpha,v,cpar)*u^alpha}.
}
\author{
Harry Joe
}

