% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb1sk-qcond.r, R/bb1sk.r
\name{qcondbb1sk21}
\alias{qcondbb1sk21}
\alias{qcondbb1sk12}
\alias{qcondbb1sk}
\alias{pbb1sk}
\alias{pcondbb1sk21}
\alias{pcondbb1sk}
\alias{pcondbb1sk12}
\alias{dbb1sk}
\title{Skew BB1 Copula Family}
\usage{
qcondbb1sk21(p, u, cparsk, eps = 1e-08, mxiter = 30, iprint = F)

qcondbb1sk12(p, v, cparsk, eps = 1e-08, mxiter = 30, iprint = F)

qcondbb1sk(p, u, cparsk, eps = 1e-08, mxiter = 30, iprint = F)

pbb1sk(u, v, cparsk)

pcondbb1sk21(v, u, cparsk)

pcondbb1sk(v, u, cparsk)

pcondbb1sk12(u, v, cparsk)

dbb1sk(u, v, cparsk)
}
\arguments{
\item{cparsk}{copula parameter vector, third component in [0,1]}

\item{eps}{tolerance for convergence}

\item{mxiter}{maximum number of Newton-Raphson iterations}
}
\description{
Adds a skew parameter alpha in [0,1]. Distribution function is found by
\code{pbb1(u/u^alpha,v,cpar)*u^alpha}.
}
\author{
Vincenzo Coia, Harry Joe
}
