% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb1skp-qcond.R, R/bb1skp.R
\name{qcondbb1skp21}
\alias{qcondbb1skp21}
\alias{qcondbb1skp12}
\alias{qcondbb1skp}
\alias{pbb1skp}
\alias{pcondbb1skp21}
\alias{pcondbb1skp}
\alias{pcondbb1skp12}
\alias{dbb1skp}
\title{Skew BB1skp Copula Family}
\usage{
qcondbb1skp21(p, u, cparsk, eps = 1e-08, mxiter = 30, iprint = F)

qcondbb1skp12(p, v, cparsk, eps = 1e-08, mxiter = 30, iprint = F)

qcondbb1skp(p, u, cparsk, eps = 1e-08, mxiter = 30, iprint = F)

pbb1skp(u, v, cparsk)

pcondbb1skp21(v, u, cparsk)

pcondbb1skp(v, u, cparsk)

pcondbb1skp12(u, v, cparsk)

dbb1skp(u, v, cparsk)
}
\arguments{
\item{cparsk}{copula parameter vector, third component in [0,1]}

\item{eps}{tolerance for convergence}

\item{mxiter}{maximum number of Newton-Raphson iterations}
}
\description{
"p" stands for permutation (u and v are switched from BB1sk).
Adds a skew parameter alpha in [0,1]. Distribution function is found by
\code{pbb1v(u/u^alpha,v,cpar)*u^alpha}.
}
\author{
Harry Joe; Vincenzo Coia
}
