% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb1vsk-qcond.R, R/bb1vsk.R
\name{qcondbb1vsk21}
\alias{qcondbb1vsk21}
\alias{qcondbb1vsk12}
\alias{qcondbb1vsk}
\alias{pbb1vsk}
\alias{pcondbb1vsk21}
\alias{pcondbb1vsk}
\alias{pcondbb1vsk12}
\alias{dbb1vsk}
\title{Skew BB1v Copula Family}
\usage{
qcondbb1vsk21(p, u, cparsk, eps = 1e-08, mxiter = 30, iprint = F)

qcondbb1vsk12(p, v, cparsk, eps = 1e-08, mxiter = 30, iprint = F)

qcondbb1vsk(p, u, cparsk, eps = 1e-08, mxiter = 30, iprint = F)

pbb1vsk(u, v, cparsk)

pcondbb1vsk21(v, u, cparsk)

pcondbb1vsk(v, u, cparsk)

pcondbb1vsk12(u, v, cparsk)

dbb1vsk(u, v, cparsk)
}
\arguments{
\item{cparsk}{copula parameter vector, third component in [0,1]}

\item{eps}{tolerance for convergence}

\item{mxiter}{maximum number of Newton-Raphson iterations}
}
\description{
Adds a skew parameter alpha in [0,1]. Distribution function is found by
\code{pbb1v(u/u^alpha,v,cpar)*u^alpha}.
}
\author{
Harry Joe; Vincenzo Coia
}
