% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cop-bb6_supp.R
\name{logdbb6}
\alias{dbb6r}
\alias{dbb6u}
\alias{dbb6v}
\alias{logdbb6}
\alias{logdbb6r}
\alias{logdbb6u}
\alias{logdbb6v}
\alias{pbb6r}
\alias{pbb6u}
\alias{pbb6v}
\alias{pcondbb6r}
\alias{pcondbb6u}
\alias{pcondbb6v}
\alias{qcondbb6r}
\alias{qcondbb6u}
\alias{qcondbb6v}
\alias{rbb6r}
\alias{rbb6u}
\alias{rbb6v}
\title{BB6 copula: supplemental functions}
\usage{
logdbb6(u, v, cpar)

pbb6r(u, v, cpar)

dbb6r(u, v, cpar)

logdbb6r(u, v, cpar)

rbb6r(n, cpar)

pcondbb6r(v, u, cpar)

qcondbb6r(p, u, cpar)

pbb6u(u, v, cpar)

dbb6u(u, v, cpar)

logdbb6u(u, v, cpar)

rbb6u(n, cpar)

pcondbb6u(v, u, cpar)

qcondbb6u(p, u, cpar)

pbb6v(u, v, cpar)

dbb6v(u, v, cpar)

logdbb6v(u, v, cpar)

rbb6v(n, cpar)

pcondbb6v(v, u, cpar)

qcondbb6v(p, u, cpar)
}
\description{
Appended "u" means a horizontal flip (i.e. switching u to 1-u).
Appended "v" means a vertical flip (i.e. switching v to 1-v).
Appended "r" means reflection/survival copula
}

