% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cop-bb8_supp.R
\name{logdbb8}
\alias{dbb8r}
\alias{dbb8u}
\alias{dbb8v}
\alias{logdbb8}
\alias{logdbb8r}
\alias{logdbb8u}
\alias{logdbb8v}
\alias{pbb8r}
\alias{pbb8u}
\alias{pbb8v}
\alias{pcondbb8r}
\alias{pcondbb8u}
\alias{pcondbb8v}
\alias{qcondbb8r}
\alias{qcondbb8u}
\alias{qcondbb8v}
\alias{rbb8r}
\alias{rbb8u}
\alias{rbb8v}
\title{BB8 copula: supplemental functions}
\usage{
logdbb8(u, v, cpar)

pbb8r(u, v, cpar)

dbb8r(u, v, cpar)

logdbb8r(u, v, cpar)

rbb8r(n, cpar)

pcondbb8r(v, u, cpar)

qcondbb8r(p, u, cpar)

pbb8u(u, v, cpar)

dbb8u(u, v, cpar)

logdbb8u(u, v, cpar)

rbb8u(n, cpar)

pcondbb8u(v, u, cpar)

qcondbb8u(p, u, cpar)

pbb8v(u, v, cpar)

dbb8v(u, v, cpar)

logdbb8v(u, v, cpar)

rbb8v(n, cpar)

pcondbb8v(v, u, cpar)

qcondbb8v(p, u, cpar)
}
\description{
Appended "u" means a horizontal flip (i.e. switching u to 1-u).
Appended "v" means a vertical flip (i.e. switching v to 1-v).
Appended "r" means reflection/survival copula
}

