% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bskewncop.R
\name{dbskewn}
\alias{dbskewn}
\alias{dbskewncop}
\alias{pbskewncop}
\alias{pcondbskewncop12}
\alias{pcondbskewncop21}
\alias{qcondbskewncop12}
\alias{qcondbskewncop21}
\title{Skew Normal Copula Functions}
\usage{
dbskewn(y1, y2, param)

dbskewncop(u, v, param)

pbskewncop(u, v, param)

pcondbskewncop21(v, u, param)

pcondbskewncop12(u, v, param)

qcondbskewncop21(p, u, param)

qcondbskewncop12(p, v, param)
}
\value{
\code{dbskewncop}: bivariate skew-normal copula density

\code{pbskewncop}: bivariate skew-normal copula cdf

\code{pcondbskewncop21}: conditional cdf P(V<v|U=u) of bivariate skew-normal copula

\code{pcondbskewncop12}: conditional cdf P(U<u|V=v) of bivariate skew-normal copula

\code{qcondbskewncop21}: conditional quantile C_{2|1}^{-1}(p|u)of bivariate skew-normal copula

\code{qcondbskewncop12}: conditional quantile C_{1|2}^{-1}(p|v)of bivariate skew-normal copula
}
\description{
skew-normal copula from Azzalini-DallaValle's bivariate skew-normal
u,v = values or vectors in (0,1)
param = (rho,ga1,ga2), with -1<rho<1, -1<ga1<1, -1<ga2<1
 and a constraint 1-ga1^2-ga2^2-rho^2+2*ga1*ga2*rho>0
}
\details{
param=c(rho,ga1,ga2), constraint is that
correlation matrix with these parameters should be positive definite
}
\note{
pcond21 = P(V<v|U=u) for bskewn copula with  y1=qsn(u), y2=qsn(v)

pcond12 = P(U<u|V=v) for bskewn copula with  y1=qsn(u), y2=qsn(v)

qcond21 = bskewn copula with  y1=qsn(u), y2=qsn(v)

qcond12 = bskewn copula with  y1=qsn(u), y2=qsn(v)
}

