% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop-bskewncop_supp.R
\name{pbskewncopp}
\alias{dbskewncopp}
\alias{logdbskewncopv}
\alias{pbskewncopp}
\alias{pcondbskewncop}
\alias{pcondbskewncopp}
\alias{pcondbskewncopp12}
\alias{pcondbskewncopp21}
\alias{qcondbskewncop}
\alias{qcondbskewncopp}
\alias{qcondbskewncopp12}
\alias{qcondbskewncopp21}
\title{Skew normal copula: supplement}
\usage{
pbskewncopp(u, v, cpar)

dbskewncopp(u, v, cpar)

logdbskewncopv(u, v, cpar)

pcondbskewncopp21(v, u, cpar)

qcondbskewncopp21(p, u, cpar)

pcondbskewncopp(v, u, cpar)

qcondbskewncopp(p, u, cpar)

pcondbskewncopp12(u, v, cpar)

qcondbskewncopp12(p, v, cpar)

pcondbskewncop(v, u, cpar)

qcondbskewncop(p, u, cpar)
}
\description{
Skew normal copula: supplement
}

