% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scoreqf.R
\name{calibration}
\alias{calibration}
\title{Calibration of a Forecast}
\usage{
calibration(y, yhat, tau, disp_plot = TRUE)
}
\arguments{
\item{y}{Vector of outcomes/response.}

\item{yhat}{Matrix of forecasts. Each row is a forecast corresponding to
\code{y}; each column corresponds to a quantile index.}

\item{tau}{Vector of the quantile indices forecast.}

\item{disp_plot}{Logical; should a P-P plot be output?}
}
\value{
A data frame with columns "ind" for quantile indices, and "exc"
for corresponding proportion of outcomes that exceeded the quantile
of that index.
}
\description{
Computes the proportion of quantile forecasts that are exceeded, for each
quantile specified.
}

