% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/center.rvine.R
\name{centervarray}
\alias{centervarray}
\title{Center a Vine Array}
\usage{
centervarray(A)
}
\arguments{
\item{A}{Vine array, possibly truncated.}
}
\value{
A vine array with the same dimensions as \code{A}.
}
\description{
Converts a vine array \code{A} so that the first variables (up to
truncation) are not leaves. So, a slightly weaker condition than
natural order. Deprecated; use \code{\link{center.rvine}} instead.
}
\details{
For a \code{t}-truncated vine array \code{(t < ncol(A)-1)},
the vine array is re-ordered so that the first \code{t} variables
introduced in the outputted array are not leaves.

If \code{t = ncol(A)-1}, then the entered vine isn't truncated, and the
first \code{t-1} variables in the outputted array are not leaves (in fact,
a natural order array is outputted, since it satisfies that requirement).
}

