% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cnqrobj.R
\name{cnqrobj}
\alias{cnqrobj}
\title{Make CNQR objective function}
\usage{
cnqrobj(y, yhat_fun, taus, g = identity)
}
\arguments{
\item{y}{A vector of observed responses/outcomes.}

\item{yhat_fun}{A function of the parameter vector that returns a matrix
with \code{length(y)} rows of forecasts corresponding to \code{y}.
Columns should be the quantile forecasts corresponding to \code{taus}.}

\item{taus}{A vector specifying the quantile indices that are being
predicted.}

\item{g}{A vectorized non-decreasing function for which to transform
both response/outcome and forecast in the estimation.}
}
\value{
Returns a function that takes parameter values and returns
the value of the objective function.
}
\description{
Makes the CNQR objective function for you.
}
\note{
You don't really need to use this function, since it's almost
identical to the \code{\link{scoreq}} function.
}
\seealso{
\code{\link{space_taus}}
}

