% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnstr_H.R
\name{cnstr_H}
\alias{cnstr_D1H}
\alias{cnstr_H}
\alias{cnstr_Hinv}
\title{Construction Function for the IG Copula Family}
\usage{
cnstr_H(t, theta, k)

cnstr_D1H(t, theta, k)

cnstr_Hinv(w, theta, k, ngrid = 1000, silent = TRUE)
}
\arguments{
\item{t}{Vector of values >=1 to evaluate the function at.}

\item{theta}{Value of second argument of H_{k}, >0. This is allowed
to be a vector, except in \code{cnstr_Hinv()}.}

\item{k}{Single numeric >1 corresponding to the \code{k} parameter
of the function.}

\item{w}{Vector of values in [0,1] to evaluate the inverse function at.}

\item{silent}{Logical; should the message output by \code{pcinterpolate()}
be silenced?}
}
\description{
\code{cnstr_H} is the function itself, and \code{cnstr_Hinv} is
its inverse; \code{cnstr_D1H} is the first-argument
derivative.
}

