% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnstr_Psi.R
\name{cnstr_Psi}
\alias{cnstr_D2Psi}
\alias{cnstr_DPsi}
\alias{cnstr_Psi}
\alias{cnstr_Psiinv}
\title{Construction function for the IGL copula family}
\usage{
cnstr_Psi(t, k)

cnstr_DPsi(t, k)

cnstr_D2Psi(t, k)

cnstr_Psiinv(w, k, mxiter = 20, eps = 1e-06, bd = 5)
}
\arguments{
\item{t}{Vector of values to evaluate the function at, >=0.}

\item{k}{Single numeric >1. Parameter of the function.}

\item{w}{Vector of values to evaluate the inverse function at, between
0 and 1 (inclusive)}
}
\description{
\code{cnstr_Psi} is the function itself, and \code{cnstr_Psiinv} is
its inverse; \code{cnstr_DPsi} is the
derivative; and \code{cnstr_D2Psi} is the second derivative.
}
\examples{
## Some examples of evaluating the functions.
arg <- c(0, 0.5, 3, Inf, NA)
cnstr_Psi(arg, k=2)
cnstr_DPsi(arg, k=1.2)
cnstr_DPsi(arg, k=2)
cnstr_DPsi(arg, k=3)
cnstr_Psiinv(c(0, 0.5, 1), k=1.5)

## Visual
foo <- function(u) cnstr_Psiinv(u, k=1.5)
curve(foo)
}

