% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvine.R
\name{combine_copmat}
\alias{combine_copmat}
\title{Combine Copula and Parameter Matrices}
\usage{
combine_copmat(rv, copmat, cparmat, digits = 3)
}
\arguments{
\item{rv}{Object of type "rvine"}

\item{copmat}{Matrix of copula families, as in an "rvine" object.}

\item{cparmat}{Matrix of copula parameters, as in an "rvine" object.}

\item{digits}{Number of significant digits to round parameter values to.}
}
\value{
Just see an example, but here's a description:
Character matrix with entries starting with that in \code{copmat},
followed by the corresponding parameter values in \code{cparmat}
separated by commas, housed in parentheses.
}
\description{
Combines a copula matrix and a parameter matrix, so that specific
copulas are indicated.
}
\note{
\code{copmat} and \code{cparmat} are not needed if \code{rv} is specified,
and vice-versa.
}
\examples{
G <- AtoG(CopulaModel::Dvinearray(4))

## Example 1
rv <- rvine(G, "frk", 2)
combine_copmat(rv)
combine_copmat(copmat=rv$copmat, cparmat=rv$cparmat)

## Example 2
rv <- rvine(G, "bvtcop", c(0.4, 5))
combine_copmat(rv)

## Example 3
rv <- rvine(matrix(4:1, ncol = 4))
combine_copmat(rv)
}
