% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newcop-constr_funs-limit.R
\name{cnstr_ell}
\alias{cnstr_ell}
\alias{cnstr_ellp}
\alias{cnstr_ellp2}
\title{Construction functions for the new copula}
\usage{
cnstr_ell(t, k)

cnstr_ellp(k, t)

cnstr_ellp2(k, t)
}
\description{
Construction functions for the new copula
}
\examples{
ell(0:10, k=3)
ell(0:10, k=1)
ell(0:10, k=99999)
}

