% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copnum2name.R
\name{copnum2name}
\alias{copname2num}
\alias{copnum2name}
\title{Convert between Copula Number and Name}
\usage{
copnum2name(num)

copname2num(name)
}
\arguments{
\item{num}{Vector of integer codes.}

\item{name}{Vector of copula family names (without indication of rotation)}
}
\value{
For \code{copnum2name}, returns a character vector of the copula family
names (down to the rotation). \code{NA} for non-existent families.

For \code{copname2num}, returns a list with entries corresponding to
\code{name} of the integer codes for that family, including all the rotations
if they're different. \code{NULL} entry if copula model not found.
}
\description{
The \code{VineCopula} package codes the copula families by some integer, whereas
the CopulaModel package uses a naming convention. These functions converts
the integer code to the copula family name (\code{copnum2name}), or
vice versa (\code{copname2num}).
}
\seealso{
See the help page for \code{\link{RVineCopSelect}}
in the \code{VineCopula} package for the mapping from code to copula family.
}

