% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cparspace.R
\name{cparspace}
\alias{cparspace}
\title{Get Copula Parameter Bounds}
\usage{
cparspace(cop, fn = TRUE)
}
\arguments{
\item{cop}{Single copula family name (character).}

\item{fn}{Logical; should a function of the parameter space be returned? If
not, returns a list of the lower and upper bounds.}
}
\value{
Returns a function that accepts a vector of copula
parameters (corresponding to \code{cop}) and returns \code{TRUE} if that
parameter is in the parameter space and \code{FALSE} otherwise.

In the case that you enter a copula family that is not on record in this
function, a function
that always returns \code{TRUE} is returned.
}
\description{
For a selection of bivariate copula families, returns the parameter space.
}
\note{
I didn't notice the \code{\link{cparbound}} function in the
\code{CopulaModel} package. A wrapper is still useful though, so as
to include reflected copulas (-u, -v, or -r), and to output a function
of the parameter space. Coming soon.
}
\examples{
cparspace(c("frk", "bvtcop", "bvtcop", "gum"))
cparspace("bvncop")(0.8)
cparspace("thiscopuladoesnotexist")("89")
}

