% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cparvec2cpar.R
\name{cparvec2cpar}
\alias{cparvec2cpar}
\title{Incorporate Parameter Lengths}
\usage{
cparvec2cpar(cparvec, len)
}
\arguments{
\item{cparvec}{Vector. Intended to be copula parameters.}

\item{len}{Vector of non-negative integers indicating how many parameters
belong to each family. Should have \code{sum(len)=length(cparvec)}.}
}
\value{
List of copula parameters for each family corresponding to the
entries in \code{len}.
}
\description{
Converts a vector of copula parameters, along with information about
how many copula parameters belongs to each family, into a list
of parameters for each family.
}
\examples{
cparvec2cpar(-3:3/2, len = c(1, 2, 1, 0, 2, 1, 0))
}
