% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drvine.R
\name{logdrvine}
\alias{drvine}
\alias{logdrvine}
\title{Density of a Regular Vine}
\usage{
logdrvine(dat, rv)

drvine(dat, rv)
}
\arguments{
\item{dat}{Data matrix. Rows are observations, and columns are variables.
Could be a vector if there's only one observation.}

\item{rv}{Regular vine object}
}
\value{
Vector of length = number of observations in \code{dat}, representing
the evaluated joint density of the variables in \code{A}.
}
\description{
Evaluates the density of a regular vine model (\code{drvine}) or log density
(\code{logdrvine}).
}
\details{
This function is a wrapper for
\code{rvinellkv.trunc2} in the \code{CopulaModel} package.
}
\examples{
set.seed(123)
A <- truncvarray(CopulaModel::Cvinearray(4), 2)
copmat <- makeuppertri(c("gum", "gal", "bvtcop",
                         "bvncop", "frk"), row = 2, col = 4, blanks = "")
cparmat <- makeuppertri.list(c(1.5, 1.5, 0.9, 3, 0.1, 0.5),
                             len = c(1,1,2,1,1), row = 2, col = 4)
rv <- rvine(A=A, copmat=copmat, cparmat=cparmat)
dat <- fvinesim(10, rv)
logdrvine(dat, rv)
drvine(runif(4), rv)

## The variables in A don't need to refer to all data:
u <- runif(4)
drvine(u, subset(rv, 3:4))
## ...is the same as:
drvine(u[3:4]), subset(rv, 3:4)
}
\seealso{
\code{\link{rvine}}
}

