% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evi.R
\name{evi.Hill}
\alias{evi.Hill}
\title{Hill's estimator based on largest observations}
\usage{
evi.Hill(sample, k, SEmethod = "norm")
}
\arguments{
\item{sample}{Numeric vector representing the univariate sample.}

\item{k}{Integer in [1,n] where \code{n=length(sample)}.}

\item{SEmethod}{Method used to estimate the standard error of the
estimate. If standard error is not required, put \code{NULL}. See details.}
}
\value{
if \code{SEmethod=NULL}, returns a single numeric estimate of
the tail index xi. Otherwise, returns a 1x2 data frame of headers
\code{estimate} and \code{SE}.
}
\description{
From a univariate sample, returns the Hill estimate based on
the largest k+1 observations.
}
\details{
The methods to estimate the standard error of the estimate is
currently \code{'norm'}, which uses the asymptotic Gaussian approximation
evaluated at the estimate of the tail index.
}
\examples{
evi.Hill(rcauchy(1000), 100)
evi.Hill(rcauchy(1000), 100, NULL)
## Can't compute if some values are non-positive:
evi.Hill(rnorm(1000) - 500, 100)
}
\seealso{
See the package 'smoothtail' for potentially more estimates.
}

