% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evi.R
\name{evi.Pickands}
\alias{evi.Pickands}
\title{Pickands Estimator based on largest observations}
\usage{
evi.Pickands(sample, four_k, SEmethod = "norm")
}
\arguments{
\item{sample}{Numeric vector representing the univariate sample.}

\item{four_k}{Integer in [1,(n-1)] where \code{n=length(sample)}.}

\item{SEmethod}{Method used to estimate the standard error of the
estimate. If standard error is not required, put \code{NULL}. See details.}
}
\value{
if \code{SEmethod=NULL}, returns a single numeric estimate of
the tail index xi. Otherwise, returns a 1x2 data frame of headers
\code{estimate} and \code{SE}.
}
\description{
Computes the Pickands estimate of the extreme value index based
on a univariate sample, using the largest \code{4k+1} observations.
}
\details{
The methods to estimate the standard error of the estimate is
currently \code{'norm'}, which uses the asymptotic Gaussian approximation
evaluated at the estimate of the tail index.
If \code{four_k} is not a multiple of 4, it will be reduced to
the nearest multiple of 4 that is less than \code{four_k}.
}
\examples{
evi.Pickands(runif(1000), 50)
evi.Pickands(rcauchy(1000), 50)
evi.Pickands(runif(1000), 50, NULL)
}
\seealso{
See the package 'smoothtail' for potentially more estimates.
}

