% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit.BN.R
\name{fit.BN}
\alias{fit.BN}
\title{Fit a Bayesian Network model}
\usage{
fit.BN(y, xdat, ymarg = identity, xmargs = identity, xord = NULL,
  cops = NULL, families = c("bvncop", "bvtcop", "mtcj", "gum", "frk", "joe",
  "bb1", "bb7", "bb8"), ...)
}
\arguments{
\item{y}{Vector of response observations.}

\item{xdat}{Matrix of predictors. Columns are variables, rows are observations
(that correspond with entries of \code{y}).}

\item{ymarg}{Vectorized function of the cdf of the response.}

\item{xmargs}{List of vectorized functions of the cdfs of the univariate
predictors, in the order of the columns of \code{xdat}. Or, a single function
if they're all the same.}

\item{xord}{If you already know what order you want to pair the response
and the predictors, put that order here as a vector of "column numbers".}

\item{cops}{If you already know some of the copula models that you'd like
to use corresponding to the variables in \code{xord}, put them here.
Entries with \code{NA} will have a copula model chosen.}

\item{families}{A vector of copula family names to try
fitting (will also consider their rotations/reflections). Limited to
those families available in \code{VineCopula} package, listed in
\code{\link{BiCopSelect}}.}

\item{...}{Other arguments to pass to \code{VineCopula::BiCopSelect}.}
}
\value{
A list with three entries:

\itemize{
     \item \code{$xord}: The pairing order, as a vector of the column numbers
     of \code{xdat}.
     \item \code{$cops}: A vector of copula family names corresponding to
     the pairs in the order in \code{$xord}.
     \item \code{$cparstart}: A list of copula parameters corresponding to
     the families in \code{$cops}, which should be taken as starting values
     for an optimization (obtained by bivariate likelihood). If each entry
     of the list is a vector of length one, then a vector of those values
     is returned here instead.
}
}
\description{
Finds copula models for each pair of response and predictor in some order,
so that the pair is conditional on previous predictors.
}
\details{
This function first determines the order to pair up the response
and predictors in the order of highest partial correlation (by using
\code{\link{lm}}). Then the bivariate copula models are chosen and fitted
individually using \code{VineCopula::BiCopSelect}.

By "pairing response and predictors in some order", I mean pairing (Y,X1),
(Y,X2)|X1, (Y,X3)|(X1,X2), etc, though not necessarily in that order.

For the \code{familyset} argument, the default is almost all of
the families available. It just doesn't include the Tawn copula families.
}
\examples{
## Get some simulated data:
set.seed(6277)
p <- 5
A0 <- CopulaModel::Dvinearray(p)
copmat0 <- makeuppertri("frk", p-1, p, "")
cparmat0 <- makeuppertri(3, p-1, p)
dat <- fvinesim(100, A0, copmat0, cparmat0)
y <- dat[, 1]
xdat <- dat[, -1]

## Fit the model:
fit.BN(y, xdat)
}

