% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit.rvine.R
\name{fit.rvine}
\alias{fit.rvine}
\title{Fit a Vine Model}
\usage{
fit.rvine(xdat, vars = 1:ncol(xdat), ntrunc = ncol(xdat) - 1,
  margs = identity, A = NULL, families = c("bvncop", "bvtcop", "mtcj",
  "gum", "frk", "joe", "bb1", "bb7", "bb8"), ...)
}
\arguments{
\item{xdat}{Matrix of data; columns represent variables, and rows observations.}

\item{vars}{Vector of integers specifying the column numbers of \code{xdat}
to fit a model to. Default is all variables.}

\item{ntrunc}{Integer, either \code{1, 2, ...,ncol(xdat)-1},
of the truncation level of the vine to be fit.}

\item{margs}{List of vectorized functions of the univariate marginal cdf's of
the data, in the order of the columns. Or a single function if the cdf's
are all the same.}

\item{A}{If you know the vine array that you want to use, put it here.
Labels should correspond to column numbers of \code{xdat}. Otherwise,
leave it \code{NULL}.}

\item{families}{A vector of copula family names to try
fitting (will also consider their rotations/reflections). Limited to
those families available in \code{VineCopula} package, listed in
\code{\link{BiCopSelect}}.}

\item{...}{Other arguments to pass to \code{VineCopula::RVineCopSelect}.}
}
\value{
A list with three entries:

\itemize{
     \item \code{$cdf}: List of length \code{ncol(xdat)} of the marginal
     distribution functions, as input in \code{margs}.
     \item \code{$A}: Vine array, truncated to \code{ntrunc}.
     \item \code{$copmat}: \code{ntrunc x ncol(A)} upper-triangular
     matrix of copula model names.
     \item \code{$cparmat}: \code{ntrunc x ncol(A)} upper-triangular
     matrix of copula parameters. If at least one copula family has more or
     less parameters than 1, each entry is a list of length one containing
     the vector of copula parameters for that copula family.
}

Note that the \code{$cdf} output is listed so that the position of the cdf
in the list corresponds to the integer labels in \code{$A}.
This might change in the future if it's more convenient or sensible
to only listing the cdfs for variables in \code{$A}.
}
\description{
Fits a joint distribution for the data using a vine copula. The vine array
is first chosen using the minimum spanning tree algorithm using
the function \code{gausstrvine.mst} in the \code{CopulaModel}
package, then the pairwise
copula models are chosen and fit using \code{RVineCopSelect} in the
\code{VineCopula} package.
}
\details{
For the \code{familyset} argument, the default is almost all of
the families available. It just doesn't include the Tawn copula families.

Note that you'll need the \code{igraph} package installed.
}
\note{
Because of some conflict between the deprecated package \code{igraph0}
(used by package \code{CopulaModel})
and the newer package \code{igraph} (used by package \code{VineCopula}),
this function can only be run a maximum of one time.
Run it more and you'll get an error.
}
\examples{
## Get some simulated data:
set.seed(152)
ntrunc <- 2
p <- 4
A0 <- truncvarray(CopulaModel::Dvinearray(p), ntrunc)
copmat0 <- makeuppertri("frk", ntrunc, p, "")
cparmat0 <- makeuppertri(3, ntrunc, p)
dat <- fvinesim(100, A0, copmat0, cparmat0)

## Fit a model to the data:
fit.rvine(dat, ntrunc=ntrunc)
fit.rvine(dat, c(4, 2, 3))
}

