% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitbicop.R
\name{fitbicop_lh}
\alias{fitbicop_lh}
\alias{fitbicop_qcond}
\title{Fit a Bivariate Copula Model}
\usage{
fitbicop_lh(u, v, families = c("indepcop", "bvncop", "bvtcop", "mtcj", "gum",
  "frk", "joe", "bb1", "bb7", "bb8"), cpar = NULL, method = "aic")

fitbicop_qcond(u, v, families = c("indepcop", "bvncop", "bvtcop", "mtcj",
  "gum", "frk", "joe", "bb1", "bb7", "bb8"), cpar = NULL, taus = 1:10/11 *
  0.1 + 0.9)
}
\arguments{
\item{u}{Vector of Unif(0,1) data.}

\item{v}{Vector of Unif(0,1) data, same length as \code{u}.}

\item{families}{Vector of copula families to try fitting. Rotations
of each family are automatically considered.}

\item{method}{Either "aic" or "bic" (case doesn't matter)}

\item{taus}{Vector of quantile indices to score.}
}
\value{
A list with entries \code{$cop} (the best-fitting copula family),
\code{$cpar} (the fitted copula parameters), and \code{$aic} and \code{$bic}.
}
\description{
These functions choose and fit the best copula model for data with
uniform margins. Intended for internal use. \code{fitbicop_qcond}
not done yet.
}
\note{
This function is not that smart -- it assumes not all cpars are
specified, and if some are specified, that the copula model is known.

I'm using BiCopSelect() just because it has smart starting points.
}
\examples{
library(CopulaModel)
set.seed(13)
u <- runif(100)
v <- qcondbvtcop(runif(100), u, c(0.9, 1))
fitbicop_lh(u, v)
fitbicop_lh(u, v, families = "bvtcop", cpar = c(NA, NA))

v <- qcondgum(runif(100), u, 4)
fitbicop_lh(u, 1-v)

v <- runif(100)
fitbicop_lh(u, v)
fitbicop_lh(u, v, families = "indepcop")
}

