% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitbicop.R
\name{fitbicop_cnqr}
\alias{fitbicop_cnqr}
\title{Fit bicop}
\usage{
fitbicop_cnqr(u, v, uval, vval, QYhat, families = c("indepcop", "bvncop",
  "bvtcop", "mtcj", "gum", "frk", "joe", "bb1", "bb7", "bb8"), cpar = NULL,
  taus = 1:10/11 * 0.1 + 0.9)
}
\arguments{
\item{QYhat}{function of the previous}

\item{cpar}{list of copula parmeters you wish to fit to each family in \code{families}.
\code{NULL} if you don't want to specify any; otherwise, copulas for which
you don't want to specify should have a \code{NULL} entry; otherwise,
put a vector of at least partially specified parameters, with \code{NA}
going in place of parameters in that family that you don't know about.}

\item{taus}{Vector of quantile indices to score.}
}
\description{
Fit bicop
}

