% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit.rvine.R
\name{fitrvine_basic}
\alias{fitrvine_basic}
\title{Fit a Vine to Data}
\usage{
fitrvine_basic(dat, vars = 1:ncol(dat), ntrunc = length(vars) - 1,
  families = c("bvncop", "bvtcop", "mtcj", "gum", "frk", "joe", "bb1", "bb7",
  "bb8"))
}
\arguments{
\item{dat}{Data matrix with uniform scores.}

\item{vars}{Vector of variables (column numbers in \code{dat}) to fit a model to.}

\item{ntrunc}{Truncation level of vine.}

\item{families}{Copula families to fit. Sorry, no 'indepcop'.}
}
\description{
Fits a vine to data using \code{\link{gausstrvine.mst}} and then
fitting the vine array layer-by-layer using \code{\link{fitrvine_basic}}.
Only fits a regular vine, not a generalized vine.
}
\examples{
set.seed(123)
dat <- matrix(runif(500), ncol = 5)
fitrvine_basic(dat)
fitrvine_basic(dat, 3)
fitrvine_basic(dat, integer(0))
}

