% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitseq.R
\name{fitseq}
\alias{fitseq}
\alias{fitseq_rvine}
\title{Choose sequence of variables to link}
\usage{
fitseq(dat, var, linkwith = (1:ncol(dat))[-var], a = NULL)

fitseq_rvine(dat, var, linkwith = (1:ncol(dat))[-var], a = NULL, A)
}
\arguments{
\item{dat}{Matrix of data with Unif(0,1) margins}

\item{var}{Integer; the variable you want to link.}

\item{linkwith}{Integer vector; the variables you want to link up \code{var}
with.}

\item{a}{Vector of a pre-specified vine array column, if you know
some of it. Should start with \code{var}, and have \code{NA} for blank spots.}

\item{A}{Vine array matrix containing variables in \code{linkwith} and
\code{a[-1]} (to ensure the order is chosen to preserve the rvine status).}
}
\value{
Vector of the variable sequence, with \code{var} first.
}
\description{
Chooses the sequence of variables that link up with another variable
in the order of highest to lowest partial correlation, as approximated by
lm(). Intended to internal use. Use \code{fitseq_rvine} if you want
the order to be chosen so that, when attached to a base rvine, you end
up with an rvine still (NOT FUNCTIONAL).
}
\details{
All variables in \code{linkwith} will be linked, despite what's in \code{a}.
}
\examples{
set.seed(123)
dat <- matrix(runif(50), ncol=5)
fitseq(dat, 1, 3:5)
fitseq(dat, 1, 2:5, a=c(1,NA,4))
}

