% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitvarray.R
\name{fitvarray}
\alias{fitvarray}
\title{Fit copula models to a vine array.}
\usage{
fitvarray(dat, A, copmat = NULL, cparmat = NULL, families = c("indepcop",
  "bvncop", "bvtcop", "mtcj", "gum", "frk", "joe", "bb1", "bb7", "bb8"))
}
\arguments{
\item{dat}{Data matrix with Uniform margins.}

\item{A}{Vine array matrix, possibly truncated.}

\item{copmat}{Pre-specified copula families in the form of an upper-triangular
matrix. Put \code{NA} to leave the edge unspecified. \code{NULL} for
fully unspecified.}

\item{cparmat}{Pre-specified copula parameters corresponding to some of the
specified copulas in \code{copmat}. Put \code{NA} in place of parameters to
leave them unspecified. \code{NULL} for fully unspecified.}

\item{famililes}{Vector of candidate copula family names.}
}
\description{
Choose and fit copula models on a pre-specified vine array. Intended for
internal use, as opposed to \code{RVineCopSelect}. Currently still
uses \code{VineCopula} through \code{BiCopSelect}.
}
\note{
Expecting smart input. So, don't input a vine array that has no edges,
and don't specify a parameter for an unspecified copula family.
}

