% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igcop_helper.R
\name{get_grid_helper}
\alias{get_grid_helper}
\title{Evaluate the helper function at a grid}
\usage{
get_grid_helper(w, theta, k, ngrid = 1000)
}
\arguments{
\item{w}{Vector of values in (0,1) for which the grid should cover. Should not
contain \code{NA}s.}

\item{theta}{Single numeric >0; parameter of \code{\link{igcop_helper}}}

\item{k}{Single numeric >1; parameter of \code{\link{igcop_helper}}}

\item{ngrid}{Positive integer; how many points at a time should the
interpolation grid be expanded by?}
}
\value{
\code{get_grid_helper} returns a matrix with
2 columns: column one contains values of \code{t}, and column two
contains \code{igcop_helper} evaluated at those \code{t} values (which
should "engulf" \code{clean_w}).
}
\description{
Chooses a grid on the real line so that evaluating the function
\code{\link{igcop_helper}} at those points ends up spanning its range (0,1)
and encapsulating \code{w}. It's used so that
\code{\link{igcop_helper}} can be inverted by interpolation.
}
\details{
This function begins by obtaining a first guess at a grid whose
evaluated \code{\link{igcop_helper}} would encapsulate \code{w}. It
does this by inverting a rudimentary approximation function of
\code{\link{igcop_helper}},  1-exp(-theta/k*exp(t)) for t real. Then
if need be,
the grid is expanded by adding more points, chosen
by linear extrapolation based on the slopes
at either end of the grid.
}
\note{
This algorithm is somewhat rudimentary. It would be nice to upgrade it
so that it finds a more "focussed" grid that doesn't necessarily have to
span the entire range of the \code{\link{igcop_helper}} function, but just
the range of \code{w}. But if \code{w} is too narrow, the algorithm
is susceptible for being too sparse for \code{\link{igcop_helper}}
to be interpolated accurately. So \code{w} is currently expanded if it's not
"wide enough".
}
\examples{
## Grid for the theta=k=3 function:
foo <- get_grid_helper(0.4, 3, 3)
plot(foo[, 1], foo[, 2])

foo <- get_grid_helper(1:99/100, 3, 3) # Same thing.
plot(foo[, 1], foo[, 2])

## Here's one where the grid needed expanding:
foo <- get_grid_helper(0.4, 0.1, 1.1)
nrow(foo)
plot(foo[, 1], foo[, 2])
}

