% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cop-gum_supp.R
\name{pgumu}
\alias{dgumu}
\alias{dgumv}
\alias{logdgumr}
\alias{logdgumu}
\alias{logdgumv}
\alias{pcondgumu}
\alias{pcondgumv}
\alias{pgumu}
\alias{pgumv}
\alias{qcondgumu}
\alias{qcondgumv}
\alias{rgumu}
\alias{rgumv}
\title{Gumbel copula: supplemental functions}
\usage{
pgumu(u, v, cpar)

dgumu(u, v, cpar)

logdgumu(u, v, cpar)

rgumu(n, cpar)

pcondgumu(v, u, cpar)

qcondgumu(p, u, cpar)

pgumv(u, v, cpar)

dgumv(u, v, cpar)

logdgumv(u, v, cpar)

rgumv(n, cpar)

pcondgumv(v, u, cpar)

qcondgumv(p, u, cpar)

logdgumr(u, v, cpar)
}
\description{
Appended "u" means a horizontal flip (i.e. switching u to 1-u).
Appended "v" means a vertical flip (i.e. switching v to 1-v).
Appended "r" means reflection/survival copula
}

