% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncop-cnstr_Phi.R
\name{igamma}
\alias{igamma}
\title{Upper Incomplete Gamma Function}
\usage{
igamma(k, t)
}
\arguments{
\item{k}{Indicated exponent in the integrand. Positive.}

\item{t}{Non-negative values of the lower limit of the integral.}
}
\description{
The upper incomplete gamma function, defined as the integral
integrate(x^{k-1} exp(-x), x=t..infinity).
}
\note{
This function is vectorized over either argument, but not both at once.
}

