% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igcop_helper.R
\name{igcop_helper}
\alias{igcop_helper}
\alias{igcop_helper_inv}
\title{IG Copula helper function}
\usage{
igcop_helper(t, theta, k)

igcop_helper_inv(w, theta, k, ngrid = 1000, silent = TRUE)
}
\arguments{
\item{t}{Numeric vector to evaluate \code{igcop_helper} at. Real.}

\item{theta}{Single numeric >0.}

\item{k}{Single numeric >1.}

\item{w}{Numeric vector to evaluate the inverse function at,
\code{igcop_helper_inv}. In [0,1].}

\item{silent}{Logical; should the message output by \code{pcinterpolate()}
be silenced?}
}
\value{
\code{igcop_helper} and \code{igcop_helper_inv} both return
vectors of length equal to the length of their first argument, representing
the function evaluations.
}
\description{
These functions are useful for computing the 2|1
distribution of the IG copula family. We have
C_{2|1}(v|u) = igcop_helper(log(log(x)), theta*(1-u), k), where
x=cnstr_Hinv(1-v, theta, k),
so that the quantile function can be written in terms of
\code{igcop_helper_inv}. The inverse is found using
interpolation, for which \code{\link{get_grid_helper}} finds a grid of
\code{t} values such that \code{igcop_helper} evaluates to points
that contain the \code{w}'s input into \code{igcop_helper_inv}.
}
\note{
Since an interpolation is used to invert \code{igcop_helper},
which is a different function for each \code{theta},
\code{igcop_helper_inv} only allows one theta.
}

