% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop-ig_supp.R
\name{pigcopu}
\alias{pigcopu}
\alias{digcopu}
\alias{logdigcopu}
\alias{rigcopu}
\alias{pcondigcopu}
\alias{qcondigcopu}
\alias{pigcopv}
\alias{digcopv}
\alias{logdigcopv}
\alias{pcondigcopv}
\alias{qcondigcopv}
\alias{pcondigcopr}
\alias{qcondigcopr}
\alias{pcondigcopr12}
\alias{qcondigcopr12}
\alias{digcopr}
\alias{pigcopr}
\alias{rigcopr}
\title{IG copula: supplemental functions}
\usage{
pigcopu(u, v, cpar)

digcopu(u, v, cpar)

logdigcopu(u, v, cpar)

rigcopu(n, cpar)

pcondigcopu(v, u, cpar)

qcondigcopu(p, u, cpar)

pigcopv(u, v, cpar)

digcopv(u, v, cpar)

logdigcopv(u, v, cpar)

pcondigcopv(v, u, cpar)

qcondigcopv(p, u, cpar)

pcondigcopr(v, u, cpar)

qcondigcopr(p, u, cpar)

pcondigcopr12(u, v, cpar)

qcondigcopr12(p, v, cpar)

digcopr(u, v, cpar)

pigcopr(u, v, cpar)

rigcopr(n, cpar)
}
\description{
Appended "u" means a horizontal flip (i.e. switching u to 1-u).
Appended "v" means a vertical flip (i.e. switching v to 1-v).
Appended "r" means reflection/survival copula
}
