% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iglcop.R
\name{qcondiglcop}
\alias{diglcop}
\alias{pcondiglcop}
\alias{pcondiglcop12}
\alias{pcondiglcop21}
\alias{piglcop}
\alias{qcondiglcop}
\alias{qcondiglcop12}
\alias{qcondiglcop21}
\title{IGL Copula Family Functions}
\usage{
qcondiglcop(tau, u, cpar)

pcondiglcop(v, u, cpar)

qcondiglcop21(tau, u, cpar)

pcondiglcop21(v, u, cpar)

pcondiglcop12(u, v, cpar)

qcondiglcop12(tau, v, cpar)

diglcop(u, v, cpar)

piglcop(u, v, cpar)
}
\arguments{
\item{tau}{Vector of quantile levels in [0,1] to evaluate a quantile function
at.}

\item{u, v}{Vectors of values in [0,1] representing values of the first
and second copula variables.}

\item{cpar}{Single numeric >1; corresponds to parameter \code{k} in the
IGL copula family.}
}
\value{
Numeric vector of length equal to the length of the input vector(s).
}
\description{
Functions related to the IGL copula family, denoted  by \code{'iglcop'}.
}
\note{
Inputting two vectors greater than length 1 is allowed, if they're
the same length.
Also, \code{qcondiglcop21} and \code{pcondiglcop21} are the same as
\code{qcondiglcop} and \code{pcondiglcop} -- their the distributions of
variable 2 given 1.
}

