% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indepcop.R
\name{pcondindepcop}
\alias{pcondindepcop}
\alias{qcondindep}
\alias{qcondindepcop}
\title{Conditional Independence Copula}
\usage{
pcondindepcop(v, u, cpar = 0)

qcondindepcop(tau, u, cpar = 0)

qcondindep(tau, u, cpar = 0)
}
\description{
These are just the identity functions with respect to the first argument.
}

