% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varray_helper_functions.R
\name{invert.perm}
\alias{invert.perm}
\title{Invert a permutation}
\usage{
invert.perm(perm)
}
\arguments{
\item{perm}{Vector of integers in \code{{1:length(perm)}}}
}
\value{
A vector of length \code{length(perm)} of the inverse permutation.
}
\description{
For a permutation of a set of integers \code{1, 2, ...,p}, finds the
inverse permutation using the \code{\link{which}} function.
}
\note{
This function won't check whether the integers you input are
in the set \code{{1:length(perm)}}, but allows for length-0 entry.
}
\examples{
perm <- c(5, 1, 2, 3, 4)
(perminv <- invert.perm(perm))
perm[perminv]
perminv[perm]

## The zero case:
invert.perm(integer(0))
}
