% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcondseq.R
\name{is.dvine}
\alias{is.dvine}
\title{Is a Vine Array a D-Vine?}
\usage{
is.dvine(A)
}
\arguments{
\item{A}{Vine array matrix.}
}
\value{
Logical -- \code{TRUE} if \code{A} is a D-vine. \code{FALSE} if not.
If \code{A} is not a matrix, or has \code{integer(0)} columns,
it'll return \code{NULL}.
}
\description{
Is a Vine Array a D-Vine?
}
\examples{
is.dvine(CopulaModel::Dvinearray(5))
is.dvine(CopulaModel::Cvinearray(6))
is.dvine("hello")
}

