% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcondseq.R
\name{is.vineleaf}
\alias{is.vineleaf}
\title{Are variables Vine Leaves?}
\usage{
is.vineleaf(A, select = diag(A))
}
\arguments{
\item{A}{Vine array matrix.}

\item{select}{Vector of vine variables that you want to query.}
}
\value{
Vector of length \code{length(select)} of logicals.
}
\description{
Check whether some variables on a vine are "leaves" -- that is, only
connect to one other variable on the first tree.
}
\examples{
A <- CopulaModel::Dvinearray(5)
is.vineleaf(A, 5)
A <- rvinesubset(A, 2:4)
is.vineleaf(A, 3:4)
is.vineleaf(A, integer(0))
}

