% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop-joe_supp.R
\name{pjoer}
\alias{pjoer}
\alias{djoer}
\alias{logdjoer}
\alias{rjoer}
\alias{pcondjoer}
\alias{qcondjoer}
\alias{pjoeu}
\alias{djoeu}
\alias{logdjoeu}
\alias{rjoeu}
\alias{pcondjoeu}
\alias{qcondjoeu}
\alias{pjoev}
\alias{djoev}
\alias{logdjoev}
\alias{rjoev}
\alias{pcondjoev}
\alias{qcondjoev}
\title{Joe copula: supplemental functions}
\usage{
pjoer(u, v, cpar)

djoer(u, v, cpar)

logdjoer(u, v, cpar)

rjoer(n, cpar)

pcondjoer(v, u, cpar)

qcondjoer(p, u, cpar)

pjoeu(u, v, cpar)

djoeu(u, v, cpar)

logdjoeu(u, v, cpar)

rjoeu(n, cpar)

pcondjoeu(v, u, cpar)

qcondjoeu(p, u, cpar)

pjoev(u, v, cpar)

djoev(u, v, cpar)

logdjoev(u, v, cpar)

rjoev(n, cpar)

pcondjoev(v, u, cpar)

qcondjoev(p, u, cpar)
}
\description{
Appended "u" means a horizontal flip (i.e. switching u to 1-u).
Appended "v" means a vertical flip (i.e. switching v to 1-v).
Appended "r" means reflection/survival copula
}
