% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/layeropts.R
\name{layeropts}
\alias{layeropts}
\title{Fitting options for a new Vine Layer}
\usage{
layeropts(var = NULL, A = NULL, ntrunc = NULL, cops = NULL,
  cpar = NULL, families = c("bvncop", "bvtcop", "mtcj", "gum", "frk", "joe",
  "bb1", "bb7", "bb8"))
}
\arguments{
\item{ntrunc}{Truncation level. Could be a vector corresponding to the
truncation level for the variables \code{var} or \code{A[1, ]}.}
}
\description{
When fitting a new layer(s) to a vine, use this function to specify
"known" components of the new layer(s), as well as
}
\note{
The arrays here use the newer form, where variables go in row 1.
}

