% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logdR.R
\name{logdR}
\alias{logdR}
\title{Density of a Regular Vine}
\usage{
logdR(dat, A, copmat, cparmat, Fmarg = identity)
}
\arguments{
\item{dat}{Data matrix. Rows are observations, and columns are variables.
Could be a vector if there's only one observation.}

\item{A}{Vine array. Integer labels should correspond to the column number
in \code{dat}.}

\item{copmat}{Matrix of bivariate copula model names, corresponding to
the upper-triangular part of \code{A}.}

\item{cparmat}{Matrix of parameters for copula models in \code{copmat}.
If a copula model has more or less parameters than 1, put the vector of
copula parameters in a list in each entry.}

\item{Fmarg}{List of vectorized marginal cdfs of data, in the order listed
in \code{dat}. Or a single such function if they're all the same.}
}
\description{
Density of a Regular Vine
}

