% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{marginal}
\alias{marginal}
\title{Estimate Marginal distribution}
\usage{
marginal(x, dist.name = NULL, ecdf.split = NULL, init.val = NULL,
  in.range = NULL, soften.cdf = TRUE, eqf.type = 1)
}
\arguments{
\item{x}{Vector representing the univariate sample drawn from the
distribution to be estimated.}

\item{dist.name}{String; name of the parametric family of distributions, if
you don't want the edf. Use R's naming conventions
for distributions ahead of the d/p/q/r. For example, the Normal
distribution would be \code{"norm"}.}

\item{ecdf.split}{Numeric; real number for which to fit the specified model
for data above such number, and empirical distribution to be fit below.}

\item{init.val}{If a parametric distribution is specified (through
\code{dist.name}), this should be the initial value specification for the
parameters. Could be (1) a vector, in which the order provided will correspond
to the order given in the R documentation for that distribution, or (2) a list
with names being the parameter names, or (3) remain \code{NULL} in which
case default parameter values are used (if there even are defaults).}

\item{in.range}{The same as the argument in \code{\link{rnlm}} -- a function
that takes the parameter values and returns \code{TRUE} if that value of
the parameter is "proper"; \code{FALSE} otherwise.}

\item{soften.cdf}{Logical; should the cdf be modified so that it never returns
the boundary cases 0 or 1? (uses 0.5/n or 1-0.5/n instead,
where n is amount of data)}

\item{eqf.type}{If the empirical quantile function is being estimated, this
argument specifies the type of quantile algorithm to use, as in the
\code{type} argument of the \code{\link{quantile}} function. In particular,
an integer between 1 and 9.}
}
\value{
If \code{return.param = FALSE} (default), output is a list of
length three:

\code{$cdf}: The estimated distribution function

\code{$qf}: The estimated quantile function (inverse of the cdf estimate)

\code{$pdf}: The estimated density function or mass function. Gives a kernel
density estimate if using the empirical distribution.
}
\description{
Estimates the marginal distribution of a univariate sample, either
by MLE (if a parametric family is specified) or empirical distribution (edf,
with kernel density pdf)
otherwise.
}
\details{
Some noteworthy options for \code{eqf.type} are: 1 = inverse
empirical cdf (default of \code{marginal}); 7 = default of
\code{\link{quantile}}, a continuous function.
}

