% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop-mtcj_supp.R
\name{pmtcju}
\alias{dmtcju}
\alias{dmtcjv}
\alias{logdmtcju}
\alias{logdmtcjv}
\alias{pcondmtcju}
\alias{pcondmtcjv}
\alias{pmtcju}
\alias{pmtcjv}
\alias{qcondmtcju}
\alias{qcondmtcjv}
\alias{rmtcju}
\alias{rmtcjv}
\title{MTCJ copula: supplemental functions}
\usage{
pmtcju(u, v, cpar)

dmtcju(u, v, cpar)

logdmtcju(u, v, cpar)

rmtcju(n, cpar)

pcondmtcju(v, u, cpar)

qcondmtcju(p, u, cpar)

pmtcjv(u, v, cpar)

dmtcjv(u, v, cpar)

logdmtcjv(u, v, cpar)

rmtcjv(n, cpar)

pcondmtcjv(v, u, cpar)

qcondmtcjv(p, u, cpar)
}
\description{
Appended "u" means a horizontal flip (i.e. switching u to 1-u).
Appended "v" means a vertical flip (i.e. switching v to 1-v).
Appended "r" means reflection/survival copula
}

