% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dvine.R
\name{pcondD}
\alias{pcondD}
\alias{qcondD}
\title{Conditional D-Vine Distribution}
\usage{
pcondD(y, x, copmat, cparmat, FXmarg = identity, FYmarg = identity)

qcondD(tau, x, copmat, cparmat, FXmarg = identity, QY = identity)
}
\arguments{
\item{y}{response(s) to evaluate the cdf at. Could be a vector (or matrix
with one column) to evaluate at multiple cdf's corresponding to the rows
of \code{x}.}

\item{x}{vector of covariates X_1,...,X_p; or, a matrix with different
covariate values (p columns).}

\item{copmat}{Upper triangular \code{ntrunc x p} matrix of
copula names (like "frk" and "gum") as strings in the D-vine array.
So, rows correspond to tree
depth in the vine; columns correspond to the edges in that tree (in the order
of the D-vine, i.e. "left to right").}

\item{cparmat}{Upper triangular matrix of copula parameters corresponding
to the copula families in \code{copmat}.}

\item{FXmarg}{List of (univariate) marginal cdfs of X_1, ..., X_p;
each should be vectorized. Or a single function if the cdf is all the same.}

\item{FYmarg}{Marginal cdf of Y, vectorsized.}

\item{tau}{Vector of quantile indices to evaluate conditional quantile
functions}

\item{QY}{Quantile function of Y, vectorized.}
}
\description{
For the vine Y--X_1--...--X_p,
these functions return the joint cdf or
quantile function of Y|X_1,...,X_p.
}
\details{
The copulas link (downstream variable, upstream variable), not
the other way around. This doesn't make a difference if the copula is "U-V"
symmetric.
}
\note{
You might want to load the CopulaModel package to load the
copula families.
}
\examples{
## Joint distribution info
library(CopulaModel)
copmat <- makeuppertri("frk", 5, 5)
cparmat <- makeuppertri(1:10, 5, 5)

## cdf of X1|X2,...,X5 at -2|-1:2
pcondD(-2, -1:2, copmat, cparmat, pnorm)
}

