% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/newcopula.R
\name{pcondnew}
\alias{pcondnew}
\alias{pcondnew12}
\alias{pcondnew21}
\alias{qcondnew}
\title{Conditional distributions of bivariate new copula (\code{\link{pnew}}).}
\usage{
pcondnew(v, u, cpar)

pcondnew21(v, u, cpar)

qcondnew(tau, u, cpar)

pcondnew12(u, v, cpar)
}
\arguments{
\item{v}{value in interval 0,1; could be a vector}

\item{u}{value in interval 0,1; could be a vector}

\item{cpar}{copula parameter (single number >= 0)}

\item{tau}{quantile indices in [0,1]; could be a vector.}
}
\description{
\code{pcondnew} and \code{pcondnew21} are the same functions -- they're
the cdf of V|U. \code{pcondnew12} is the cdf of U|V. \code{qcondnew} is
the quantile function of V|U ('U|V' not supported at this point, since
we only really care about V|U).
}
\details{
The first two arguments should have the same length. If not,
one of them should have length 1.
}
\note{
\code{pcondnew21} is only included for completeness, since
\code{pcondnew12} is introduced as well (being a non-symmetric copula).
}

