% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcondseq.R
\name{pcondseq.generic}
\alias{pcondseq.generic}
\title{Find sequential conditional cdfs}
\usage{
pcondseq.generic(ord, dat, fX, Fcond = NULL)
}
\arguments{
\item{ord}{Integer vector; variables in the order that you'll be linking them
up with the response (so we'll find \code{ord[1]}, \code{ord[2]|ord[1]}, etc.).}

\item{dat}{\code{p}-length vector, or \code{p}-columns matrix of predictors.}

\item{fX}{Function; the density of the covariates. Should accept a vector
with each component in the space (-Inf, Inf) and return a non-negative real.}

\item{Fcond}{If you already have some of the conditional distributions,
put them here in a list to speed up algorithm (either as a
function (see details), or a vector already
evaluated at the data). Make a \code{NULL} entry if you don't have that cdf.
The \code{k}th entry should correspond to the cdf of
\code{ord[k]|ord[1:(k-1)]}.}
}
\value{
If \code{dat} is a vector, returns a
vector of evaluated cdfs of predictors
\code{ord[1]}, \code{ord[2]|ord[1]}, ..., \code{ord[p]|ord[1:(p-1)]}.

If \code{dat} is a matrix, returns a matrix of such evaluated cdfs.
}
\description{
From a \code{p}-variate joint distribution,
finds the conditional cdfs of variables \code{1}, \code{2|1}, ...,
\code{p|1:(p-1)} evaluated at some data.
This function is intended as a preliminary step before connecting a response
to covariate \code{1}, then \code{2|1}, ...,\code{p|1:(p-1)}.
Allows for permutations of \code{1:p} too.
}
\details{
If you include a function as an entry in \code{Fcond}, it should
accept a vector representing variables \code{ord[c(k, 1:(k-1))]}.
}
\note{
If some of your covariates don't have support on (-Inf, Inf), be sure
that the density still evaluates properly (to zero) outside of the support,
because this function integrates from -Inf to Inf.
}
\examples{
(dat <- matrix(rnorm(3*5), ncol = 3))
pdf <- function(x) prod(dnorm(x))
pcondseq.generic(c(3, 1), dat, fX=pdf, Fcond = list(pnorm, NULL))
}

