% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/newcopula.R
\name{phi}
\alias{phi}
\title{Construction ('phi') function for the new copula}
\usage{
phi(theta, t)
}
\arguments{
\item{theta}{Real number for the "shape" parameter of the function. Can't
be a vector.}

\item{t}{Vector of real numbers >= 1 to evaluate the function at.}
}
\description{
phi[theta](t) = (1-t^(-theta)) / (t*theta*log(t))
}

