% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncop.R, R/newcop-constr_funs.R
\name{phiinv}
\alias{phiinv}
\title{Inverse of the construction function}
\usage{
phiinv(w, theta, k, mxiter = 20, eps = 1e-06, bd = 5)

phiinv(w, theta, k, mxiter = 20, eps = 1e-06, bd = 5)
}
\arguments{
\item{w}{Vector of values in in (0,1] to find the inverse at.}

\item{theta}{Non-negative index for the construction function.}

\item{theta}{Non-negative index for the construction function.}

\item{w}{Vector of values in in (0,1] to find the inverse at.}
}
\description{
Uses 'newer' version suggested by Harry. It's possible that this
function could get stuck in an infinite loop, but it's never happened
before.

Uses 'newer' version suggested by Harry. It's possible that this
function could get stuck in an infinite loop, but it's never happened
before.
}

