% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncop.R
\name{phiinvold}
\alias{phiinvold}
\title{Inverse of the construction function -- old version with uniroot.}
\usage{
phiinvold(theta, w)
}
\arguments{
\item{theta}{Real number for the "shape" parameter of the function. Can't
be a vector.}

\item{w}{Number in (0,1] to find the inverse at. Cannot be a vector.}
}
\description{
Inverse of function 'phi'.
}

