% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/newcopula.R
\name{phip}
\alias{phip}
\title{Derivative of the construction function (with respect to argument)}
\usage{
phip(theta, t)
}
\arguments{
\item{theta}{Non-negative index for the phi function.}

\item{t}{Numeric values >=1 to evaluate the function at. Could be a vector.}
}
\description{
diff(phi[theta](t), t)
}

