% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plotcop_qcurve}
\alias{plotcop_qcurve}
\title{Plot Copula Fits}
\usage{
plotcop_qcurve(u, v, cops, cpars, tauset = space_taus(10))
}
\arguments{
\item{u}{Uniform data}

\item{v}{Uniform data}

\item{cops}{Vector of copula families to show fit for.}

\item{cpars}{List of copula parameters for the families in \code{cops}.}

\item{tauset}{Vector of quantile indices to plot.}
}
\description{
Plots normal scores plots, with fitted quantile curves.
}
\examples{
set.seed(1234)
## Get data
u <- runif(100)
v <- CopulaModel::qcondgum(runif(100), u, 4)

## Fit models:
cops <- c("frk", "gum", "joe", "mtcj")
cpars <- lapply(cops, function(cop_)
fitbicop_lh(u, v, families=cop_)$cpar)

## Get plots:
plotcop_qcurve(u, v, cops, cpars)

## How about on a validation set?
u <- runif(100)
v <- CopulaModel::qcondgum(runif(100), u, 4)
plotcop_qcurve(u, v, cops, cpars)

## Over the entire range of quantile functions?
##  You can add ggplot layers too:
library(ggplot2)
plotcop_qcurve(u, v, cops, cpars, tauset = 1:19/20) +
  labs(title = "Normal Scores Plot", x = "x", y = "y")
}

