% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plotcop_score}
\alias{plotcop_score}
\title{Plot CNQR scores}
\usage{
plotcop_score(cops, scores)
}
\arguments{
\item{cops}{Vector of names of the models.}

\item{scores}{Vector of numeric scores of the models.}
}
\value{
A plot (\code{ggplot} command) of the models' scores
in descending order. The labels are "copula" on the x-axis, and
"CNQR score" on the y-axis, but you can change that by literally adding
a layer \code{labs(x="Your label here", y="Your label here")} to the output.
}
\description{
Plots scores of different model in descending order for easy comparison.
}
\examples{
set.seed(12)
plotcop_score(LETTERS[1:5], rexp(5))
}

