% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncop-general.R, R/ncop.R
\name{pnew}
\alias{dnew}
\alias{logdnew}
\alias{pnew}
\title{Bivariate new copula -- cdf and density}
\usage{
pnew(u, v, cpar)

logdnew(u, v, cpar)

dnew(u, v, cpar)

pnew(u, v, cpar)

logdnew(u, v, cpar)

dnew(u, v, cpar)
}
\arguments{
\item{u}{value in interval 0,1; could be a vector}

\item{v}{value in interval 0,1; could be a vector}

\item{cpar}{copula parameters (theta>=0, k>=1).}

\item{u}{value in interval 0,1; could be a vector}

\item{v}{value in interval 0,1; could be a vector}

\item{cpar}{copula parameter (single number >= 0)}
}
\description{
\code{pnew} is the distribution function; \code{dnew} is the density;
\code{logdnew} is the log of the density. This version is the 2-parameter
version, with cpar=c(theta, k).

#' @rdname pcondnew
#' @export
qcondnew <- function(tau, u, cpar){
    if (length(tau) == 1 & length(u) > 1){
        tau <- rep(tau, length(u))
    }
    if (length(tau) > 1 & length(u) == 1){
        u <- rep(u, length(tau))
    }
    if (length(tau) != length(u)){
        return(NA)
    }
    len <- length(tau)
    if (cpar[1] == 0){
        return(tau)
    } else {
        return(sapply(1:len, function(i)
            cnstr(qcondnew_helper(tau[i], cpar[1]*u[i], cpar[2]), cpar)))
    }
}

\code{pnew} is the distribution function; \code{dnew} is the density;
\code{logdnew} is the log of the density.
}
\details{
Either at least one of \code{u} or \code{v} should have length=1, or
\code{u} and \code{v} should have the same length.

Either at least one of \code{u} or \code{v} should have length=1, or
\code{u} and \code{v} should have the same length.
}

