% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_appendy.R
\name{predict_appendy}
\alias{predict_appendy}
\title{Quantile Model Predictions}
\usage{
predict_appendy(appendy, dat = "val", tau = space_taus(10))
}
\arguments{
\item{appendy}{A fitted object with \code{\link{fitlayer_cnqr}}.}

\item{dat}{Either a data matrix to make predictions on, or one of the
two strings: \code{"val"} will make predictions on the validation set, and
\code{"tr"} will make predictions on the training set.}

\item{tau}{Vector of quantile indices to make predictions at.}
}
\description{
Quantile Model Predictions
}

