% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qcondBN.R
\name{qcondBN}
\alias{qcondBN}
\alias{qcondBNu}
\title{Conditional Distribution: Bayesian Network Method}
\usage{
qcondBN(tau, x, p, cop, cpar, fX, Fcond = NULL, QY)

qcondBNu(tau, u, p, cop, cpar, fU, Fcond = NULL, QY)
}
\arguments{
\item{tau}{Vector of quantile indices to evaluate at.}

\item{x}{Vector of predictor values \code{X_1, ..., X_p}. Length should
equal the number of variables conditioning on.}

\item{p}{Integer; number of variables \code{X_1,...,X_p} to condition on.}

\item{cop}{Vector of copula names (like "frk"), whose \code{i}th entry
links \code{(X_i, Y)|X_1,...,X_{i-1}}.}

\item{cpar}{List of copula parameters corresponding to \code{cop}. Or,
vector of parameters if each copula family in \code{cop} has one parameter.}

\item{fX}{Function that accepts a vector argument; the
density of \code{(X_1,...,X_p)}.}

\item{Fcond}{If you have the conditional cdf
\code{X_i|X_1,...,X_{i-1}} for some \code{i}'s, insert them here
in a list (in their respective components). Put \code{NULL}
for entries that you don't have the form for. You could also already
evaluate this function at \code{x} if you'd like.}

\item{QY}{Quantile function for the response.}
}
\value{
Vector of conditional quantile functions evaluated at the quantile
indices \code{tau}.
}
\description{
Finds the quantile function of a response \code{Y} given predictors
\code{X_1,...,X_p} through a Bayesian Network. The bivariate distributions
\code{(X_i, Y)|X_1,...,X_{i-1}} are
modelled using bivariate copula
models. The distribution of the response is assumed known/fitted, along
with the joint distribution of the predictors.
}
\details{
If \code{Fcond} is not missing any cdf's, then there's no
need to specify \code{fX}.
}

