% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncop-general.R
\name{qcondnew_helper}
\alias{qcondnew_helper}
\title{Helper function for qcondnew}
\usage{
qcondnew_helper(w, a, k, mxiter = 20, eps = 1e-06, bd = 5)
}
\arguments{
\item{w}{Value to evaluate the inverse at. Should be in [0,1] (with 0
returning \code{Inf}). Vectorized.}

\item{a, k}{Parameters, as above.}
}
\description{
This is the inverse of the function that maps t to
(1-pgamma(a*log(t),k-1))/t, which has domain [1, infinity) and range
(0,1].
}

