% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QYgX.R
\name{QYgX}
\alias{QYgX}
\title{Quantile Function for Y|X}
\usage{
QYgX(tau, ucond, cops, cpars, QYfitted)
}
\arguments{
\item{tau}{Vector of quantile indices to evaluate at. Or,
different vectors (of the same length) can be specified
for each observation by making this a matrix.}

\item{ucond}{Matrix of uniform "partial" predictors. Columns are ordered
by linkage with the response; rows are observations.}

\item{cops}{Vector of copula model names, corresponding to the columns of
\code{ucond}.}

\item{cpars}{List of parameter vectors corresponding to \code{cops}.}

\item{QYfitted}{Vectorized fitted (marginal) quantile
function of the response.}
}
\value{
A matrix, with columns being the quantile levels, and rows
corresponding to the observations in \code{ucond}.
}
\description{
Computes the conditional quantile function of a response
at specified quantile indices. The model is obtained by specifying
copulas that link together the response with a sequence of
"partial" predictors (conditional on the previous predictors). Intended
for internal use.
}
\examples{
(dat <- matrix(runif(3*6), ncol = 3))
tau <- c(0.1, 0.3, 0.7, 0.9)
cops <- c("gum", "bvtcop", "frk")
cpars <- list(3.5, c(0.6, 3), 2.5)
QYgX(tau, dat, cops = cops, cpars = cpars, QYfitted = qexp)
}

