% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/relabel.rvine.R
\name{relabel.rvine}
\alias{relabel.rvine}
\alias{relabelvarray}
\title{Relabel Variables in a Vine/Vine Array}
\usage{
\method{relabel}{rvine}(rv, labs = 1:ncol(rv$G))

relabelvarray(G, labs = 1:ncol(G))
}
\arguments{
\item{rv}{Object of type 'rvine'}

\item{labs}{Vector of new labels. The order of the labels correspond to
the order of the variables in the vine array \code{G}.}

\item{G}{Vine array}
}
\value{
The inputted \code{rv} or \code{G} with the array matrix
relabelled.
}
\description{
Relabel Variables in a Vine/Vine Array
}
\details{
Similar to \code{CopulaModel::varrayperm} but allows for
the posibility that \code{G} is not square, as well as labels outside
of the set \code{{1:ncol(G)}}.
}
\examples{
(G <- AtoG(CopulaModel::Cvinearray(5))[1:3, ])
rv <- rvine(G, "frk", 4)

relabelvarray(G, c(3, 2, 1, 5, 4))
relabel(rv, c(3, 2, 1, 5, 4))

## Labels don't need to be in the set {1:ncol(G)}.
relabelvarray(G, c(54, 234, 1, 35, 42))
relabel(rv, c(54, 234, 1, 35, 42))
}

