% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/relabel.rvine.R
\name{relabel.rvine}
\alias{relabel.rvine}
\title{Relabel Variables in a Regular Vine}
\usage{
\method{relabel}{rvine}(rv, labs = 1:ncol(rv$A))
}
\arguments{
\item{rv}{Object of type 'rvine'.}

\item{labs}{Vector of new labels. The order of the labels correspond to
the order of the variables in the vine array \code{A}.}
}
\value{
An object of type 'rvine' with a relabelled vine array matrix.
}
\description{
Relabel Variables in a Regular Vine
}
\details{
Similar to \code{CopulaModel::varrayperm} but allows for
non-square vine arrays, as well as labels outside
of the set \code{{1:ncol(rv$A)}}.
}
\examples{
(rv <- rvine(truncvarray(CopulaModel::Cvinearray(5), 2)))
relabel(rv, c(3, 2, 1, 5, 4))

## Labels don't need to be in the set {1:ncol(A)}.
A <- CopulaModel::Dvinearray(4)
A <- apply(A, 1:2, function(i) if (i==0) "" else letters[i])
relabel.varray(rvine(A), c("this", "is", "a", "demo"))
relabel.varray(rvine(A))
}

