% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/varray_helper_functions.R
\name{relabel.varray}
\alias{relabel.varray}
\title{Relabel Variables in a Vine Array}
\usage{
relabel.varray(A, labs = 1:ncol(A))
}
\arguments{
\item{A}{Vine array. Could be truncated.}

\item{labs}{Vector of new labels. The order of the labels correspond to
the order of the variables in the vine array \code{A}.}
}
\value{
A relabelled vine array matrix.
}
\description{
Relabel Variables in a Vine Array
}
\details{
Identical to \code{CopulaModel::varrayperm} but allows for
the posibility that \code{A} is not square.
}
\examples{
(A <- truncvarray(CopulaModel::Cvinearray(5), 2))
relabel.varray(A, c(3, 2, 1, 5, 4))
}

