% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/releaf.rvine.R
\name{releafvarray}
\alias{releaf.rvine}
\alias{releafvarray}
\title{"Re-leaf" a Vine}
\usage{
releafvarray(G, leaf)

\method{releaf}{rvine}(rv, leaf)
}
\arguments{
\item{G}{Regular vine array.}

\item{leaf}{Vector of variables that you want to
get new vine array for, if such a vine array exists.}

\item{rv}{Object of type "rvine".}
}
\value{
A vine array or vine with \code{leaf} at the end, if such a
vine array exists; \code{NULL} otherwise.
}
\description{
Convert a vine array (\code{releafvarray}) or a vine
(\code{releaf.rvine}) so that a variable of your choice appears last
in the vine array -- if possible. This might only work for
regular vines, with arrays truncated in the traditional sense.
}
\examples{
A <- makeuppertri(c(4,4,4,4,6,5,
                    6,6,6,4,4,
                    1,1,1,6,
                    5,5,1,
                    2,2,
                    3), 6, 6, incDiag=T)
G <- AtoG(A)
releafvarray(G, 2)
releafvarray(G, 1)
releafvarray(truncvarray(G, 0), 1)
releaf(rvine(G, "frk", 3), 2)
}

