% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/releaf.varray.R
\name{releaf.rvine}
\alias{releaf.rvine}
\title{"Re-leaf" a Vine Array}
\usage{
\method{releaf}{rvine}(rv, leaf)
}
\arguments{
\item{rv}{Regular vine object.}

\item{leaf}{Integer; variable you want to make a leaf of the array.}
}
\value{
A regular vine object with \code{leaf} at the end
of the vine array, if such a
vine exists; \code{NULL} otherwise.
}
\description{
Convert a vine array so that a variable(s) of your choice appears last
(bottom-right corner) of the vine array -- if possible.
}
\examples{
A <- makeuppertri(c(4,4,4,4,6,5,
                    6,6,6,4,4,
                    1,1,1,6,
                    5,5,1,
                    2,2,
                    3), 6, 6, incDiag=T)
rv <- rvine(A)
releaf(rv, leaf = 2)
releaf(rv, leaf = 1)
releaf(trunc(rv, 2), leaf = 1)

## 1:ncol(A) not required as variables.
rv <- subset(rv, vars(rv)[1:4])
releaf(A, leaf = 1)
}

