% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/releaf.varray.R
\name{releaf.varray}
\alias{releaf.varray}
\title{"Re-leaf" a Vine Array}
\usage{
releaf.varray(A, leaf = varray.vars(A))
}
\arguments{
\item{A}{Vine array matrix, possibly truncated.}

\item{leaf}{Vector of variables that you want to
get new vine array for, if such a vine array exists.}
}
\value{
A vine array (matrix) with \code{leaf} at the end, if such a
vine array exists; \code{NULL} otherwise. If \code{length(leaf) > 1},
a list of such output is returned, corresponding to the entries
in \code{leaf}.
}
\description{
Convert a vine array so that a variable(s) of your choice appears last
(bottom-right corner) of the vine array -- if possible.
}
\examples{
A <- makeuppertri(c(4,4,4,4,6,5,
                    6,6,6,4,4,
                    1,1,1,6,
                    5,5,1,
                    2,2,
                    3), 6, 6, incDiag=T)
releaf.varray(truncvarray(A, 2))
releaf.varray(truncvarray(A, 3), leaf = 3:6)
releaf.varray(A, leaf = 5)

## The function doesn't require 1:ncol(A) as variables.
A <- truncvarray(CopulaModel::Dvinearray(5), 2)
A <- apply(A, 1:2, function(i) if (i==0) "" else letters[i])
releaf.varray(A)
}

