% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.dvine_is.vineleaf.R
\name{releafvarray}
\alias{releafvarray}
\title{"Re-leaf" a Vine Array}
\usage{
releafvarray(A, ntrunc = ncol(A) - 1, leaves = diag(A))
}
\arguments{
\item{A}{Vine array matrix.}

\item{ntrunc}{Truncation integer, between 1 and \code{ncol(A)-1}.}

\item{leaves}{Vector of variables in \code{diag(A)} that you want to
get new vine array for, if such a vine array exists.}
}
\value{
A vine array (matrix) with \code{leaves} at the end, if such a
vine array exists; \code{NULL} otherwise. If \code{length(leaves) > 1},
a list of such output is returned, corresponding to the entries
in \code{leaves}
}
\description{
Convert a vine array so that a variable(s) of your choice appears last
(bottom-right corner) of the vine array -- if possible.
}
\note{
When the truncation is less than \code{ncol(A) - 2}, the array matrices
that are returned are not \code{ncol(A) x ncol(A)} -- they're
\code{(ntrunc+1) x ncol(A)}, achieved by collapsing the variables on
the diagonal upwards.
}
\examples{
A <- makeuppertri(c(1,1,1,1,2,4,
                    2,2,2,1,1,
                    3,3,3,2,
                    4,4,3,
                    5,5,
                    6), 7, 7)[1:6, 2:7]
releafvarray(A, ntrunc = 2)
releafvarray(A, ntrunc = 3, leaves = 3:6)
releafvarray(A, leaves = 5)
}

