% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{reordervarray}
\alias{reorder.rvine}
\alias{reordervarray}
\title{Reorder a Vine, If Possible}
\usage{
reordervarray(G, ord)

\method{reorder}{rvine}(rv, ord)
}
\arguments{
\item{G}{Vine array matrix}

\item{ord}{Vector consisting of the variables in \code{G} or \code{rv},
in the order you want them to be in.}

\item{rv}{Object of type "rvine"}
}
\value{
Returns \code{NULL} if the vine cannot be re-ordered in the
requested manner. Otherwise, the re-ordered vine array (\code{reordervarray})
or "rvine" object (\code{reorder.rvine}) is returned.
}
\description{
These functions reorder a vine according to a user-specified order.
\code{reordervarray} reorders a vine array only, whereas
\code{reorder.rvine} reorders an "rvine" object.
}
\examples{
## reordervarray:
G <- AtoG(CopulaModel::Dvinearray(5))
reordervarray(G, 5:1)
reordervarray(truncvarray(G, c(0, 1, 2, 1, 3)), c(3, 2, 4, 1, 5))
reordervarray(G, c(5, 1, 4, 2, 3)) # NULL
reordervarray(G, c(5:1, 87)) # Error


## reorder.rvine:
copmat <- makevinemat("a", c("b","c"), letters[4:6], letters[7:10], zerocol=T)
cparmat <- makevinemat(1, 2:3, 4:6, 7:10, zerocol=T)
rv <- rvine(G, copmat, cparmat)
reorder(rv, 5:1)
reorder(rv, c(3, 2, 4, 1, 5))
reorder(rv, c(5, 1, 4, 2, 3)) # NULL
reorder(rv, c(5:1, 87)) # Error

## Trivial cases:
reordervarray(matrix(0,1,0), integer(0))
reorder(rvine(matrix(5)), 5)
}

