% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rho.R
\name{rho}
\alias{rho}
\title{Asymmetric absolute deviation function}
\usage{
rho(tau, u)
}
\arguments{
\item{tau}{Number in [0,1] representing the probability/order
of the quantile.}

\item{u}{Value to evaluate the loss function at. Vectorised.}
}
\description{
The loss function that defines the tau^th-quantile;
rho(tau, u) = (tau - I(u<0))*u. Vectorised over \code{u}.
}

