% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addlayer.R
\name{rmvlayer.rvine}
\alias{rmvlayer.rvine}
\title{Remove Upper Layers in a Vine}
\usage{
\method{rmvlayer}{rvine}(rv, num = 1)
}
\arguments{
\item{rv}{Object of type "rvine"}

\item{num}{Integer; number of layers to remove.}
}
\value{
Object of type "rvine".
}
\description{
Remove Upper Layers in a Vine
}

