% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncop-general.R, R/ncop.R
\name{rnew}
\alias{rnew}
\title{Simulate from the new copula family}
\usage{
rnew(n, cpar)

rnew(n, cpar)
}
\arguments{
\item{n}{Positive integer: the number of observations to generate}

\item{cpar}{The value of the copula parameter (non-negative)}

\item{n}{Positive integer: the number of observations to generate}

\item{cpar}{The value of the copula parameter (non-negative)}
}
\value{
Returns an \code{n}x2 matrix. Each row is an independent draw
from the specified new copula.

Returns an \code{n}x2 matrix. Each row is an independent draw
from the specified new copula.
}
\description{
Simulate from the new copula family

Simulate from the new copula family
}
\note{
This function is not vectorized.

This function is not vectorized.
}
\examples{
rnew(n = 10, cpar = 4.5)
rnew(n = 10, cpar = 4.5)
}

