% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrvine.R
\name{rrvine}
\alias{rrvine}
\title{Simulate from a Regular Vine}
\usage{
rrvine(n, rv, iprint = FALSE)
}
\arguments{
\item{n}{Number of observations to generate}

\item{rv}{Regular vine object, complete.}

\item{iprint}{Logical, as in \code{\link{rvinesimvec2}}, which says
"print flag for intermediate results".}
}
\value{
Matrix with \code{n} rows and \code{max(vars(rv))} columns.
}
\description{
Simulate from a Regular Vine. Essentially a wrapper
for \code{\link{rvinesimvec2}}. NOTE: Marginals not working yet, so
data are marginally Uniform.
}
\details{
The vine variables \code{vars(rv)} are the column numbers of the outputted
data. So \code{NA} columns are introduced if necessary. That is,
if the variables are outside of the set \code{{1:length(vars(rv))}},
then the columns of the outputted data outside of the set
\code{{1:max(vars(rv))}} will be \code{NA}.
}
\note{
Like in \code{\link{rvinesimvec2}}, the copula families are assumed to be
permutation symmetric.
}
\examples{
rv <- rvine(AtoG(CopulaModel::Dvinearray(5)), "frk", makeuppertri(2, 4, 5))
set.seed(123)

## Simulate 10 observations:
rrvine(10, rv)

## Simulate 0 and 1 variables too:
rrvine(10, subset(rv, integer(0)))
rrvine(10, subset(rv, 1))

## Vine variables outside of the set 1,2,3,...?
rrvine(10, relabel(rv, c(3, 8, 1, 6, 2)))
}

