% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rvine.R
\name{rvine}
\alias{rvine}
\title{Create an object of class `rvine`}
\usage{
rvine(G, copmat, cparmat)
}
\arguments{
\item{G}{G-Vine array}

\item{copmat}{Upper triangular \code{(nrow(G)-1) x ncol(G)} matrix of copula model names.
Or, a vector of length 1 if the same copula model applies to all
edges.}

\item{cparmat}{Upper triangular \code{(nrow(G)-1) x ncol(G)} matrix of
parameters taken by the copulas in \code{copmat}. Or, a single vector
of the parameter to be applied to the entire vine.}
}
\value{
An object of class `rvine`, which is a named list of the arguments.
}
\description{
Create an object of class `rvine`
}
\examples{
## Empty vine:
rvine(matrix(integer(0)))

## Independence vine:
(rv <- rvine(matrix(4:1, ncol = 4)))
## Take a look at each component of the vine:
summary(rv)

G <- AtoG(CopulaModel::Dvinearray(4))
(rv <- rvine(G, "frk", 2))
summary(rv)
(rv <- rvine(G, "bvtcop", c(0.4, 5)))

G <- AtoG(CopulaModel::Dvinearray(4))
copmat <- makevinemat("gum",
                      c("bvtcop", "frk"),
                      c("mtcj", "frk", "indepcop"),
                      c("bvncop", "joe", "mtcj", "frk"), zerocol = TRUE)
cparmat <- makevinemat(3.1,
                       list(c(0.5, 4), 2.3),
                       list(4.2, 3.5, numeric(0)),
                       c(0.5, 2.2, 2.5, 1.6), zerocol = TRUE)
rv <- rvine(G, copmat, cparmat)
}

