% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvine2fitrvine.R
\name{rvine2fitrvine}
\alias{rvine2fitrvine}
\title{Convert \code{"rvine"} to \code{"fitrvine"}}
\usage{
rvine2fitrvine(dat, basevine)
}
\arguments{
\item{dat}{Matrix of data with Unif(0,1) margins.}

\item{basevine}{Object of type 'rvine' (completely specified).}
}
\value{
Object of type 'fitrvine'.
}
\description{
Converts a fully specified regular vine into a fitted regular vine after
specifying data. Since the original regular vine is a fully specified
distribution, there are no fitted parameters.
}

