% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjcondvine.R
\name{rvinepcond}
\alias{rvinepcond}
\title{Copula Conditional Distributions}
\usage{
rvinepcond(parvec, udat, A, ntrunc, pcondmat, np)
}
\description{
These functions compute the conditional distributions of the
last variable, given the other variables in the vine.
}
\details{
parvec = vector of parameters of pair-copulas
udat = nxd matrix with uniform scores
A = dxd vine array with 1:d on diagonal
ntrunc = truncated level, assume >=1
pcondmat = matrix of names of conditional cdfs for trees 1,...,ntrunc
 (assuming only one needed for permutation symmetric pair-copulas)
  pcondmat is empty for diagonal and lower triangle,
   and could have ntrunc rows or be empty for rows ntrunc+1 to d-1
np = dxd where np[ell,j] is size for parameter th[ell,j]
  for pair-copula in tree ell, variables j and A[ell,j]
  np=0 on and below diagonal
Output:
return C_{2|1}(u2|u1), C_{3|12}(u3|u1,u2), ...  C_{d|1..d-1}(ud|u1,...u[d-1])
[This function is modification of rvinellkv.trunc2 in CopulaModel]
}
\author{
Harry Joe
}
