% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjcondvine.R
\name{rvineqcond}
\alias{rvineqcond}
\title{n = #replications or sample size
pmat = nxd matrix of levels of u1, C_{2|1}(u_2|u_1),...,
                         C_{d|1..d-1}(u_d|u[1:(d-1)])
parvec = vector of parameters to be optimized in nllk
A = dxd vine array with 1:d on diagonal
ntrunc = truncated level, assume >=1
pcondmat = matrix of names of conditional cdfs for trees 1,...,ntrunc
 (assuming only one needed for permutation symmetric pair-copulas)
qcondmat = matrix of names of conditional quantile functions for
       trees 1,...,ntrunc
  qcondmat and pcondmat are empty for diagonal and lower triangle,
   and could have ntrunc rows or be empty for rows ntrunc+1 to d-1
np = dxd where np[ell,j] is #parameters for parameter th[ell,j]
  for pair-copula in tree ell, variables j and A[ell,j]
  np=0 on and below diagonal
iinv=T to check that this is inverse of rvinepcond()
   in this case, columns of pmat come from rvinepcond()
iinv=F, get quantiles C_{d|1:(d-1)}(p|u[1:(d-1)]) based on last column of
      pmat[,d] where u[1],..,u[d-1] have been previously converted to
      u[1], C_{2|1}(u[2]|u[1]), ... C_{d-1|1:(d-2)}(u[d-1]|u[1:(d-2)])
      via rvinepcond()}
\usage{
rvineqcond(pmat, A, ntrunc, parvec, np, qcondmat, pcondmat, iinv = F)
}
\value{
n x d matrix with values in (0,1) or
      quantile C_{d|1...d-1}(p| u[1:(d-1)])
[This function is modification of rvinesimvec2 in CopulaModel]
}
\description{
n = #replications or sample size
pmat = nxd matrix of levels of u1, C_{2|1}(u_2|u_1),...,
                         C_{d|1..d-1}(u_d|u[1:(d-1)])
parvec = vector of parameters to be optimized in nllk
A = dxd vine array with 1:d on diagonal
ntrunc = truncated level, assume >=1
pcondmat = matrix of names of conditional cdfs for trees 1,...,ntrunc
 (assuming only one needed for permutation symmetric pair-copulas)
qcondmat = matrix of names of conditional quantile functions for
       trees 1,...,ntrunc
  qcondmat and pcondmat are empty for diagonal and lower triangle,
   and could have ntrunc rows or be empty for rows ntrunc+1 to d-1
np = dxd where np[ell,j] is #parameters for parameter th[ell,j]
  for pair-copula in tree ell, variables j and A[ell,j]
  np=0 on and below diagonal
iinv=T to check that this is inverse of rvinepcond()
   in this case, columns of pmat come from rvinepcond()
iinv=F, get quantiles C_{d|1:(d-1)}(p|u[1:(d-1)]) based on last column of
      pmat[,d] where u[1],..,u[d-1] have been previously converted to
      u[1], C_{2|1}(u[2]|u[1]), ... C_{d-1|1:(d-2)}(u[d-1]|u[1:(d-2)])
      via rvinepcond()
}
\author{
Harry Joe
}

