% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rvinesubset.R
\name{rvinesubset}
\alias{rvinesubset}
\title{Subset a vine array}
\usage{
rvinesubset(A, select)
}
\arguments{
\item{A}{vine array}

\item{select}{Vector of variable indices in \code{diag(A)} to subset,
if possible.}
}
\value{
Returns a vine array of the subsetted variables, or \code{NULL} if
the subset don't form a vine.
}
\description{
Find the vine connecting a subset of variables from a bigger vine, if
it exists.
}
\details{
Just a technicality:
by saying a subset "doesn't have an existing vine", I mean that
a vine can't be formed using nodes and edges from
the original -- not that the
joint distribution of the selected variables can't be created from a vine
(so as to say, for example, that the simplifying assumption of vines
doesn't hold for this distribution).
}
\examples{
## Try a D-Vine
A <- CopulaModel::Dvinearray(5)
rvinesubset(A, c(2, 3, 4))
rvinesubset(A, c(4, 1))
## The numbering doesn't have to be 1,2,3,...
A <- makeuppertri(2 + c(1,1,1,1,4, 2,2,3,1, 3,2,3, 4,2, 5), 6, 6)[1:5, 2:6]
rvinesubset(A, 2 + c(1,3,4))
}

