% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scoreqf.R
\name{scoreq}
\alias{scoreq}
\title{Scoring functions}
\usage{
scoreq(y, yhat, tau, g = identity)
}
\arguments{
\item{y}{Vector of realizations}

\item{yhat}{Forecast matrix. Each row should be a forecast with
corresponding realization in \code{y}; columns are evaluations of the
quantile function forecast at \code{tau}.}

\item{tau}{Vector of quantile indices for which quantile function is
evaluated.}

\item{g}{Increasing univariate function to transform \code{y} and \code{yhat}
with. Vectorized. Default is the identity.}
}
\description{
Score quantile forecasts against realizations. The quantile function
forecasts are scored by evaluating the quantile function at multiple
quantile indices, instead of integrating over the functions.
}
\details{
Uses a family of proper scoring rules to score a forecast.
}

