% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc.rvine.R
\name{trunc.rvine}
\alias{trunc.rvine}
\alias{truncvarray}
\title{Truncate a Regular Vine}
\usage{
\method{trunc}{rvine}(rv, ntrunc)

truncvarray(G, ntrunc)
}
\arguments{
\item{rv}{A regular vine object.}

\item{ntrunc}{Integer; truncation level. Or, vector specifying the truncation
level of each column of the G-Vine array.}

\item{G}{A vine array.}
}
\value{
If \code{ntrunc >= nrow(G) - 1}, the original vine is
returned. Otherwise, an 'rvine' object is returned
with the truncation implemented.
}
\description{
Truncates a vine (\code{trunc.rvine}), or a vine array (\code{truncvarray}).
}
\examples{
rv <- rvine(AtoG(CopulaModel::Dvinearray(6)), "frk", 2)
trunc(rv, 3)
trunc(rv, 100)
trunc(rv, 0)

## Column-specific truncation
(rv1 <- trunc(rv, c(0, 1, 2, 1, 2, 4)))
lapply(rv1, identity)
rv2 <- trunc(rv, c(4, 4, 4, 1, 2, 4))
identical(rv1, rv2)
}

