% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trunc.rvine.R
\name{trunclevel}
\alias{trunclevel}
\title{Get Truncation Level}
\usage{
trunclevel(G, overall = FALSE)
}
\arguments{
\item{G}{Vine array.}

\item{overall}{Logical; \code{TRUE} returns the overall truncation level,
\code{FALSE} the truncation level of each column.}
}
\description{
Extract the truncation level of a vine array. Intended for internal use.
}
\examples{
G <- AtoG(CopulaModel::Dvinearray(6))
G <- truncvinemat(G, c(0, 1, 2, 1, 2, 4))
trunclevel(G)
trunclevel(G, TRUE)
}

