% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trunc.R
\name{truncvarray}
\alias{truncvarray}
\title{Truncate a Vine Array}
\usage{
truncvarray(A, ntrunc)
}
\arguments{
\item{A}{A vine array, possibly truncated.}

\item{ntrunc}{Integer; truncation level}
}
\value{
If \code{ntrunc >= nrow(A) - 1}, the original vine array is
returned. Otherwise, a truncated vine array with \code{ntrunc + 1} rows is
returned.
}
\description{
Truncates a vine array, collapsing the variables upwards. Deprecated;
please use \code{\link{trunc.rvine}} instead.
}
\note{
The variables are listed along the initial diagonal of the vine
array, then continue along the bottom row.
}
\examples{
(A <- CopulaModel::Dvinearray(6))
(A <- truncvarray(A, 3))
(A <- relabel.varray(A, c(6, 2, 4, 3, 1, 5)))
truncvarray(A, 2)
}

