% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc.rvine.R
\name{truncvinemat}
\alias{truncvinemat}
\title{Truncate a Vine Matrix}
\usage{
truncvinemat(mat, ntrunc, varray = TRUE, zero = 0)
}
\arguments{
\item{mat}{Vine matrix}

\item{ntrunc}{Integer; truncation level. Or vector for each column.}
}
\value{
A matrix, with the desired truncation.
}
\description{
Truncates a vine matrix (either the vine array, copmat, or cparmat).
Intended for internal use.
}
\examples{
(G <- AtoG(CopulaModel::Dvinearray(6)))
truncvinemat(G, 3)

copmat <- makevinemat("frk", rep("frk", 2), rep("frk", 3), rep("frk", 4), zerocol=T)
truncvinemat(copmat, c(0, 1, 0, 3, 1), varray=FALSE, zero="")
}

