% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vars.rvine.R
\name{varray.vars}
\alias{varray.vars}
\title{Extract Variables in a Vine Array}
\usage{
varray.vars(A)
}
\arguments{
\item{A}{Vine array, possibly truncated.}
}
\value{
Vector of vine variables.
}
\description{
Extract variables in a vine array, possibly truncated, in the order
of the vine array. Deprecated; use \code{\link{vars.rvine}} instead.
}
\examples{
A <- CopulaModel::Dvinearray(5)
A <- relabel.varray(A, c(5, 2, 4, 3, 1))
varray.vars(A)

A <- truncvarray(A, 2)
varray.vars(A)
}

