% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vars.rvine.R
\name{vars.rvine}
\alias{vars.rvine}
\title{Extract Variables from a Regular Vine}
\usage{
\method{vars}{rvine}(rv)
}
\arguments{
\item{rv}{A regular vine object.}
}
\value{
Vector of vine variables.
}
\description{
Extract variables from a regular vine, possibly truncated, in the order
of the vine array.
}
\examples{
rv <- rvine(AtoG(CopulaModel::Dvinearray(5)))
rv <- relabel(rv, c(5, 2, 4, 3, 1))
vars(rv)

rv <- trunc(rv, 2)
vars(rv)
}

