% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/varray_helper_functions.R
\name{Atocon}
\alias{Atocon}
\alias{contoA}
\title{Convert Vine Array to Convenient Array}
\usage{
Atocon(A)

contoA(Acon)
}
\arguments{
\item{A}{Vine array}

\item{Acon}{A convenient vine array}
}
\description{
(A "convenient" array is achieved by moving the variables in a vine array,
possibly truncated, to the top row)
}

