% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cop-bb7_supp.R
\name{logdbb7r}
\alias{dbb7u}
\alias{dbb7v}
\alias{logdbb7r}
\alias{logdbb7u}
\alias{logdbb7v}
\alias{pbb7u}
\alias{pbb7v}
\alias{pcondbb7u}
\alias{pcondbb7v}
\alias{qcondbb7u}
\alias{qcondbb7v}
\alias{rbb7r}
\alias{rbb7u}
\alias{rbb7v}
\title{BB7 copula: supplemental functions}
\usage{
logdbb7r(u, v, cpar)

rbb7r(n, cpar)

pbb7u(u, v, cpar)

dbb7u(u, v, cpar)

logdbb7u(u, v, cpar)

rbb7u(n, cpar)

pcondbb7u(v, u, cpar)

qcondbb7u(p, u, cpar)

pbb7v(u, v, cpar)

dbb7v(u, v, cpar)

logdbb7v(u, v, cpar)

rbb7v(n, cpar)

pcondbb7v(v, u, cpar)

qcondbb7v(p, u, cpar)
}
\description{
Appended "u" means a horizontal flip (i.e. switching u to 1-u).
Appended "v" means a vertical flip (i.e. switching v to 1-v).
Appended "r" means reflection/survival copula
}

