% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newcop-constr_funs.R
\name{cnstr_H}
\alias{cnstr_H}
\title{Construction ('phi') function for the new copula}
\usage{
cnstr_H(t, param)
}
\arguments{
\item{t}{Vector of real numbers >= 1 to evaluate the function at.}

\item{param}{Vector of length two: c(theta, k)}
}
\description{
H[k](t;theta) = cnstr_ell[k](theta * log(t)) / t
}
\examples{
cnstr_H(1:10, c(4,4))
cnstr_H(2:10, c(4,4))
cnstr_H(1, c(4,4))
cnstr_H(1:10, c(0,4))
cnstr_H(1:10, c(4,1))
## As long as k doesn't go to 1 before theta goes to 0, we have:
cnstr_H(1:10, c(0,1))
}

