% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newcop-constr_funs-limit.R
\name{ell}
\alias{ell}
\alias{ellp}
\alias{ellp2}
\title{Construction functions for the new copula}
\usage{
ell(t, k)

ellp(k, t)

ellp2(k, t)
}
\description{
Construction functions for the new copula
}
\examples{
ell(0:10, k=3)
ell(0:10, k=1)
ell(0:10, k=99999)
}

