% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/copnum2name.R
\name{copnum2name}
\alias{copnum2name}
\title{Convert Copula Number to Name}
\usage{
copnum2name(num)
}
\arguments{
\item{num}{Integer code, or vector of integer codes.}
}
\value{
A string of the name of the copula family.
}
\description{
The VineCopula package codes the copula families by some integer, whereas
the CopulaModel package uses a naming convention. This function converts
the integer code to the copula family name.
}
\seealso{
See the help page for \code{\link{RVineCopSelect}}
in the \code{VineCopula} package for the mapping from code to copula family.
}

