% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logdR.R
\name{logdR}
\alias{dR}
\alias{logdR}
\title{Density of a Regular Vine}
\usage{
logdR(dat, A, copmat, cparmat, Fmarg = identity)

dR(dat, A, copmat, cparmat, Fmarg = identity)
}
\arguments{
\item{dat}{Data matrix. Rows are observations, and columns are variables.
Could be a vector if there's only one observation.}

\item{A}{Vine array. Integer labels should correspond to the column number
in \code{dat}.}

\item{copmat}{Matrix of bivariate copula model names, corresponding to
the upper-triangular part of \code{A}.}

\item{cparmat}{Matrix of parameters for copula models in \code{copmat}.
If a copula model has more or less parameters than 1, put the vector of
copula parameters in a list in each entry.}

\item{Fmarg}{List of vectorized marginal cdfs of data, in the order listed
in \code{dat}. Or a single such function if they're all the same.}
}
\description{
Evaluates the density of a regular vine model (\code{dR}) or log density
(\code{logdR}).
}
\details{
This function is a wrapper for
\code{rvinellkv.trunc2} in the \code{CopulaModel} package.
}
\examples{
set.seed(123)
A <- truncvarray(CopulaModel::Cvinearray(4), 2)
copmat <- makeuppertri(c("gum", "gal", "bvtcop",
                         "bvncop", "frk"), row = 2, col = 4, blanks = "")
cparmat <- makeuppertri.list(c(1.5, 1.5, 0.9, 3, 0.1, 0.5),
                             len = c(1,1,2,1,1), row = 2, col = 4)
dat <- fvinesim(10, A, copmat, cparmat)
logdR(dat, A, copmat, cparmat)
dR(c(0.5,0.5,0.5,0.5), A, copmat, cparmat)

## The variables in A don't need to refer to all data:
A <- CopulaModel::Dvinearray(6)
A <- rvinesubset(A, 3:6)
copmat <- makeuppertri("frk", 4, 4, "")
cparmat <- makeuppertri(6:1, 4, 4)
logdR(1:6/10, A, copmat, cparmat)
## is the same as...
A <- CopulaModel::Dvinearray(4)
logdR(3:6/10, A, copmat, cparmat)
}

