% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drvine.R
\name{logdrvine}
\alias{drvine}
\alias{logdrvine}
\title{Density of a Regular Vine}
\usage{
logdrvine(dat, rvine)

drvine(dat, rvine)
}
\arguments{
\item{dat}{Data matrix. Rows are observations, and columns are variables.
Could be a vector if there's only one observation.}

\item{rvine}{Object of class 'rvine'}
}
\value{
Vector of length = number of observations in \code{dat}, representing
the evaluated joint density of the variables in \code{A}.
}
\description{
Evaluates the density of a regular vine model (\code{drvine}) or log density
(\code{logdrvine}).
}
\details{
This function is a wrapper for
\code{rvinellkv.trunc2} in the \code{CopulaModel} package.
}
\examples{
set.seed(123)
A <- truncvarray(CopulaModel::Cvinearray(4), 2)
copmat <- makeuppertri(c("gum", "gal", "bvtcop",
                         "bvncop", "frk"), row = 2, col = 4, blanks = "")
cparmat <- makeuppertri.list(c(1.5, 1.5, 0.9, 3, 0.1, 0.5),
                             len = c(1,1,2,1,1), row = 2, col = 4)
rvine <- rvine(A=A, copmat=copmat, cparmat=cparmat)
dat <- fvinesim(10, A, copmat, cparmat)
logdrvine(dat, rvine)
drvine(c(0.5,0.5,0.5,0.5), A, copmat, cparmat)

## The variables in A don't need to refer to all data:
A <- CopulaModel::Dvinearray(6)
A <- rvinesubset(A, 3:6)
copmat <- makeuppertri("frk", 4, 4, "")
cparmat <- makeuppertri(6:1, 4, 4)
logdrvine(1:6/10, A, copmat, cparmat)
## is the same as...
A <- CopulaModel::Dvinearray(4)
logdrvine(3:6/10, A, copmat, cparmat)
}
\seealso{
\code{\link{rvine}}
}

