% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit.rvine.R
\name{fit.rvine}
\alias{fit.rvine}
\title{Fit a Vine Model}
\usage{
fit.rvine(xdat, ntrunc = ncol(xdat) - 1, margs = identity,
  familyset = c(1:10, 13, 14, 16:20, 23, 24, 26:30, 33, 34, 36:40), ...)
}
\arguments{
\item{xdat}{Matrix of data; columns represent variables, and rows observations.}

\item{ntrunc}{Integer, either \code{1, 2, ...,ncol(xdat)-1},
of the truncation level of the vine to be fit.}

\item{margs}{List of vectorized functions of the univariate marginal cdf's of
the data, in the order of the columns. Or a single function if the cdf's
are all the same.}

\item{familyset}{A vector of integer codes of the copula families to try
fitting. See \code{VineCopula::RVineCopSelect} for a full list.}

\item{...}{Other arguments to pass to \code{VineCopula::RVineCopSelect}.}
}
\value{
A list with three entries:

\itemize{
     \item \code{$A}: Vine array, truncated to \code{ntrunc}.
     \item \code{$copmat}: \code{ntrunc x ncol(A)} upper-triangular
     matrix of copula model names.
     \item \code{$cparmat}: \code{ntrunc x ncol(A)} upper-triangular
     matrix of copula parameters. If at least one copula family has more or
     less parameters than 1, each entry is a list of length one containing
     the vector of copula parameters for that copula family.
}

Sorry about the matrix that gets printed whenever the function is run.
I can't seem to keep \code{VineCopula::RVineCopSelect} quiet.
}
\description{
Fits a joint distribution for the data using a vine copula. The vine array
is first chosen using the minimum spanning tree algorithm using
the function \code{gausstrvine.mst} in the \code{CopulaModel}
package, then the pairwise
copula models are chosen and fit using \code{RVineCopSelect} in the
\code{VineCopula} package.
}
\details{
For the \code{familyset} argument, the default is almost all of
the families available. It just doesn't include the Tawn copula families.

Note that you'll need the \code{igraph0} package installed.
}
\note{
Because of some conflict between the deprecated package \code{igraph0}
(used by package \code{CopulaModel})
and the newer package \code{igraph} (used by package \code{VineCopula}),
this function can only be run a maximum of one time.
Run it more and you'll get an error.
}
\examples{
## Get some simulated data:
set.seed(152)
ntrunc <- 2
p <- 4
A0 <- truncvarray(CopulaModel::Dvinearray(p), ntrunc)
copmat0 <- makeuppertri("frk", ntrunc, p, "")
cparmat0 <- makeuppertri(3, ntrunc, p)
dat <- fvinesim(100, A0, copmat0, cparmat0)

## Fit a model to the data:
fit.rvine(dat, ntrunc)
}

