% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitlayer.R
\name{fitlayer_cnqr}
\alias{fitlayer_cnqr}
\title{Temporary Title}
\usage{
fitlayer_cnqr(dats, edges, basevine, tauset, QY, stoponbest = TRUE,
  w = function(x) 1, cops = NULL, cpars = NULL, families = c("indepcop",
  "bvncop", "bvtcop", "mtcj", "gum", "frk", "joe", "bb1", "bb7", "bb8"))
}
\arguments{
\item{QY}{Vectorized function of the fitted quantile function of Y.}

\item{cops}{List with vector entries of the copula families to try fitting
for each edge.}

\item{cpars}{List with entries being the list of parameters you want to enforce
on the copula models. \code{NULL} for no restrictions; \code{NULL} entry
for no restrictions on that edge; \code{NA} in a parameter vector
means that parameter needs fitting.}

\item{families}{Vector of copula families.
For those edges in \code{cop} that don't have copula families
specially selected (i.e. have \code{NULL} entries), these families are used.}
}
\description{
Temporary Title
}

