% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/friendly_vine.R
\name{fvinesim}
\alias{fvinesim}
\title{Friendly vine simulation}
\usage{
fvinesim(n, truncA, cops, cpars, iprint = FALSE)
}
\arguments{
\item{n}{Number of observations to generate}

\item{truncA}{Vine array matrix, possibly truncated (so, #columns should
equal vine copula dimension).}

\item{cops}{Vector of strings of the copula names for each edge, in "reading
order" corresponding to \code{truncA} from top-left to bottom-right.
To have the same copula for the entire tree, just name the copula (i.e.
length 1). To have the same copula for each tree, just specify the copulas
for each tree in that order (of length = number of trees).}

\item{cpars}{List, where each entry is a vector of parameters
corresponding to the copula in \code{cops}. Optionally, if each copula
family only has one parameter, this could be a vector. Should have
\code{length(cpars)} = # of edges in the (truncated) vine.}

\item{iprint}{Logical, as in \code{\link{rvinesimvec2}}, which says
"print flag for intermediate results".}
}
\description{
A (hopefully) user-friendly function to simulate from a vine copula.
Essentially a wrapper
for \code{\link{rvinesimvec2}}. (**Doesn't yet work for truncated vines,
because I can't figure out how to do truncated vines using
\code{\link{rvinesimvec2}})
}
\details{
The vine array in \code{truncA} is made up of \code{diag(1:d)} on the diagonal,
with 0 lower triangle, and the vine array in the upper triangle; then the
bottom rows are optionally removed to truncate the vine. Here, \code{d} is
the dimension of the vine copula (\code{=ncol(truncA)}).

To name the copulas, use the names as in the CopulaModel package. For
example, "Frank" is \code{"frk"}, and "Gumbel" is \code{"gum"}.
}
\note{
Like in \code{\link{rvinesimvec2}}, the copula families are assumed to be
permutation symmetric.
}
\examples{
## Vine array:
A <- CopulaModel::Dvinearray(5)
## Simulate 10 observations with Frank copulas:
fvinesim(10, A, cops="frk", cpars=rep(2, choose(5,2)))
## Same thing, but truncate after tree 2 (doesn't work yet):
fvinesim(10, A[1:2, ], cops="frk", cpars=rep(2, 7))
}

