% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop-ig_supp.R
\name{pigcopu}
\alias{digcopu}
\alias{digcopv}
\alias{logdigcopu}
\alias{logdigcopv}
\alias{pcondigcopu}
\alias{pcondigcopv}
\alias{pigcopu}
\alias{pigcopv}
\alias{qcondigcopu}
\alias{qcondigcopv}
\alias{rigcopu}
\title{IG copula: supplemental functions}
\usage{
pigcopu(u, v, cpar)

digcopu(u, v, cpar)

logdigcopu(u, v, cpar)

rigcopu(n, cpar)

pcondigcopu(v, u, cpar)

qcondigcopu(p, u, cpar)

pigcopv(u, v, cpar)

digcopv(u, v, cpar)

logdigcopv(u, v, cpar)

pcondigcopv(v, u, cpar)

qcondigcopv(p, u, cpar)
}
\description{
Appended "u" means a horizontal flip (i.e. switching u to 1-u).
Appended "v" means a vertical flip (i.e. switching v to 1-v).
Appended "r" means reflection/survival copula
}

