% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop-igl_supp.R
\name{piglcopu}
\alias{diglcopu}
\alias{diglcopv}
\alias{logdiglcopu}
\alias{logdiglcopv}
\alias{pcondiglcopu}
\alias{pcondiglcopv}
\alias{piglcopu}
\alias{piglcopv}
\alias{qcondiglcopu}
\alias{qcondiglcopv}
\alias{riglcopu}
\title{IG copula: supplemental functions}
\usage{
piglcopu(u, v, cpar)

diglcopu(u, v, cpar)

logdiglcopu(u, v, cpar)

riglcopu(n, cpar)

pcondiglcopu(v, u, cpar)

qcondiglcopu(p, u, cpar)

piglcopv(u, v, cpar)

diglcopv(u, v, cpar)

logdiglcopv(u, v, cpar)

pcondiglcopv(v, u, cpar)

qcondiglcopv(p, u, cpar)
}
\description{
Appended "u" means a horizontal flip (i.e. switching u to 1-u).
Appended "v" means a vertical flip (i.e. switching v to 1-v).
Appended "r" means reflection/survival copula
}

