% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ncop.R
\name{ktaunew}
\alias{ktaunew}
\title{Kendall's tau of new copula family}
\usage{
ktaunew(cpar)
}
\arguments{
\item{cpar}{Vector of copula paremeters (>=0)}
}
\value{
Vector of kendall tau values for the corresponding parameter values.
}
\description{
Converts copula parameter of new copula to kendall's tau
}
\details{
A wrapper for the function \code{\link{ktau}} in CopulaModel
package, using details of the new copula.
}

