% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/friendly_vine.R
\name{makeuppertri}
\alias{makeuppertri}
\alias{makeuppertri.list}
\title{Make an upper triangular matrix}
\usage{
makeuppertri(entries, row, col, blanks = 0, byRow = TRUE)

makeuppertri.list(entries, len, row, col, blanks = list(), byRow = TRUE)
}
\arguments{
\item{entries}{Vector of entries for the upper triangular part of the matrix.}

\item{row}{Number of rows in the matrix}

\item{col}{Number of columns in the matrix}

\item{blanks}{What should go in the non-diagonal entries?}

\item{byRow}{Logical; read the \code{entries} by row (as if you're reading)
(TRUE), or read in vertically (FALSE).}

\item{len}{Vector of positive integers which specify the lengths of the
individual vectors that are pooled in \code{entries}.}
}
\description{
Upper triangular matrix making made easy.
}
\note{
Use \code{makeuppertri} to make a matrix. If you want entries to be
vectors (which would have to be an array with list entries), use
\code{makeuppertri.list}.
}
\examples{
makeuppertri(1:10, 5, 5, byRow = FALSE)
makeuppertri.list(1:12, c(1, 10, 1), 3, 3)
}

