% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcondseq.R
\name{pcondseq}
\alias{pcondseq}
\title{Find sequential predictors}
\usage{
pcondseq(dat, ord, rv)
}
\arguments{
\item{dat}{Data matrix with uniform scores.}

\item{ord}{order of variables}

\item{rv}{A regular vine, supposedly fitted to \code{dat}.}
}
\value{
Matrix with rows being observations, and columns being
X[ord[1]]; X[ord[2]]|X[ord[1]]; X[ord[3]]|X[ord[1:2]]; etc.
}
\description{
For internal use.
}
\note{
Expecting smart input: ord must be at least of length 1, with
variables in the specified rvine.
}
\examples{
## Make a vine
G <- AtoG(CopulaModel::Dvinearray(4))
rv <- rvine(G, "frk", 3)

## Generate sample
set.seed(123)
dat <- rrvine(10, rv)

## Try it out:
pcondseq(dat, c(4, 3, 2), rv)
pcondseq(dat, 1, rv)
}

