% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcondseq.R
\name{pcondseq}
\alias{pcondseq}
\title{Compute Conditional Probabilities Sequentially}
\usage{
pcondseq(dat, ord, rv)
}
\arguments{
\item{dat}{Data matrix with uniform scores.}

\item{ord}{order of variables}

\item{rv}{A regular vine, supposedly describing the joint
distribution of \code{dat}.}
}
\value{
Matrix with rows being observations in \code{dat}, and columns being
\code{U[ord[1]]}; \code{U[ord[2]]|U[ord[1]]};
\code{U[ord[3]]|U[ord[1:2]]}; etc.
}
\description{
Given a vine describing the joint distribution of some uniform data, this
function sequentially evaluates conditional distributions given previous
variables in a specified order. See details.
}
\details{
Suppose \code{Ui} denotes random variable \code{i}.
If you specify \code{ord=c(5, 2, 3, 1)}, you'll get the cdfs of
\code{U5}; \code{U2|U5}; \code{U3|(U2,U5)}; and \code{U1|(U3,U2,U5)},
evaluated at the data.

In general, this function computes:

\itemize{
     \item \code{U[ord[1]]}
     \item \code{U[ord[2]]|U[ord[1]]}
     \item \code{U[ord[3]]|U[ord[1:2]]}
     \item ...
}
}
\examples{
## Make a vine
G <- AtoG(CopulaModel::Dvinearray(4))
rv <- rvine(G, "frk", 3)

## Generate sample
set.seed(123)
dat <- rrvine(10, rv)

## Try it out:
pcondseq(dat, c(4, 3, 2), rv)
pcondseq(dat, 1, rv)
}
