% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcondseq.R
\name{pcondseq.vine}
\alias{pcondseq.vine}
\title{Temp}
\usage{
pcondseq.vine(ord, xdat, rvinefit, FX = identity)
}
\arguments{
\item{ord}{Integer vector; variables in the order that you'll be linking them
up with the response (so we'll find \code{ord[1]}, \code{ord[2]|ord[1]}, etc.).}

\item{xdat}{Vector of a single observation, or matrix of multiple observations
of the variables.}

\item{rvinefit}{The vine fit to data \code{xdat}. See details.}

\item{FX}{List of vectorized functions that are the marginals corresponding
to the columns of \code{xdat}. Or just one function if it's a common function.
Default is \code{identity} so that \code{xdat} can be uniform data if you want.}
}
\description{
Temp
}
\details{
The argument \code{rvinefit} can either be:

\enumerate{
     \item the output of \code{\link{VineCopula::RVineCopSelect}} (Version 1.6), or
     \item a list with the following named entries:
     \itemize{
         \item \code{A}: The vine array, as used in the
         package \code{\link{CopulaModel}}.
         \item \code{copmat}: An upper-triangular matrix of names of the
         bivariate copula models used in the vine.
         \item \code{cparmat}: An upper-triangular matrix of copula parameters
         to use in the corresponding copula model in \code{rvinefit$copmat}.
         Each entry should be a vector with length = the number of parameters
         for that copula model. See \code{\link{makeuppertri.list}} for help.
     }
}
}

