% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/newcopula.R
\name{phiinv}
\alias{phiinv}
\title{Inverse of the construction function}
\usage{
phiinv(theta, w, mxiter = 20, eps = 1e-06, bd = 5)
}
\arguments{
\item{theta}{Non-negative index for the construction function.}

\item{w}{Vector of values in in (0,1] to find the inverse at.}
}
\description{
Uses 'newer' version suggested by Harry. It's possible that this
function could get stuck in an infinite loop, but it's never happened
before.
}

