% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ncop.R
\name{pnew}
\alias{dnew}
\alias{logdnew}
\alias{pnew}
\title{Bivariate new copula -- cdf and density}
\usage{
pnew(u, v, cpar)

logdnew(u, v, cpar)

dnew(u, v, cpar)
}
\arguments{
\item{u}{value in interval 0,1; could be a vector}

\item{v}{value in interval 0,1; could be a vector}

\item{cpar}{copula parameter (single number >= 0)}
}
\description{
\code{pnew} is the distribution function; \code{dnew} is the density;
\code{logdnew} is the log of the density.
}
\details{
Either at least one of \code{u} or \code{v} should have length=1, or
\code{u} and \code{v} should have the same length.
}

