% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabel.varray}
\alias{relabel.varray}
\title{Relabel Variables in a Vine Array}
\usage{
\method{relabel}{varray}(A, labs = 1:ncol(A))
}
\arguments{
\item{A}{Vine array. Could be truncated.}

\item{labs}{Vector of new labels. The order of the labels correspond to
the order of the variables in the vine array \code{A}.}
}
\value{
A relabelled vine array matrix.
}
\description{
Deprecated; use \code{\link{relabel.rvine}} instead.
}
\details{
Similar to \code{CopulaModel::varrayperm} but allows for
the posibility that \code{A} is not square, as well as labels outside
of the set \code{{1:ncol(A)}}.
}
\examples{
(A <- truncvarray(CopulaModel::Cvinearray(5), 2))
relabel.varray(A, c(3, 2, 1, 5, 4))

## Labels don't need to be in the set {1:ncol(A)}.
A <- CopulaModel::Dvinearray(4)
A <- apply(A, 1:2, function(i) if (i==0) "" else letters[i])
relabel.varray(A, c("this", "is", "a", "demo"))
relabel.varray(A)
}

