% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rvine.R
\name{rvine}
\alias{rvine}
\title{Create an object of class `rvine`}
\usage{
rvine(A, copmat = NULL, cparmat = NULL, marg = NULL)
}
\arguments{
\item{A}{Vine array (matrix), possibly truncated.}

\item{copmat}{Upper triangular \code{(nrow(A)-1) x ncol(A)} matrix of copula model names.}

\item{cparmat}{Upper triangular \code{(nrow(A)-1) x ncol(A)} matrix of
parameters taken by the copulas in \code{copmat}.}

\item{marg}{List of \code{ncol(A)} vectorized marginal distribution functions
with entries corresponding to the columns in \code{A}.}
}
\value{
An object of class `rvine`, which is a named list of the arguments.
}
\description{
Create an object of class `rvine`
}
\examples{
A <- CopulaModel::Dvinearray(4)
copmat <- makeuppertri(c("gum", "bvtcop", "mtcj",
                         "frk", "indepcop",
                         "frk"), 3, 4)
cparmat <- makeuppertri.list(c(2, 0.5, 4, 2,
                               1,
                               1), len = c(1,2,1,1,0,1), nrow = 3, ncol = 4)
rvine(A, copmat, cparmat)
}

