% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trunc.R
\name{trunc.rvine}
\alias{trunc.rvine}
\title{Truncate a Regular Vine}
\usage{
\method{trunc}{rvine}(rv, ntrunc)
}
\arguments{
\item{rv}{A regular vine object.}

\item{ntrunc}{Integer; truncation level.}
}
\value{
If \code{ntrunc >= nrow(A) - 1}, the original vine is
returned. Otherwise, an 'rvine' object is returned
with the truncation implemented.
}
\description{
Truncates a regular vine.
}
\note{
The variables are listed along the initial diagonal of the vine
array, then continue along the bottom row.
}
\examples{
rv <- rvine(CopulaModel::Dvinearray(6))
trunc(rv, 3)
trunc(rv, 100)
trunc(rv, 0)
}

